Crate kitlang

Kitlang is a language built for embedding/scripting purposes

The language is inspired by rust and aims to be as similar to rust as possible with some creative liberties taken.

The structure of the project is also inspired by rust and rustc.

Re-exports

pub use kitlang_macros as macros;

Modules

ast

Defines the Abstract Syntax Tree (AST) data structures produced by the parser.

compiler

intermediate

Houses the different levels of IR used in the compiler’s architecture, as well as resolution and type checking.

interpreter

Contains the HIR interpreter, as well as implementations and management of native functions for binding.

lexer

The lexer module transforms source code strings into streams of Tokens.

parser

The parser module transforms a stream of tokens into an Abstract Syntax Tree.

prelude

Public re-exports of commonly used items from the Kitlang compiler crate.

profiling

Small utilities module for profiling and displaying the execution time of code blocks in an easily read format.

spanned_error

Provides utilities for building formatted error messages with source code context and highlights.

token

Defines the type of tokens and keywords used in the Kitlang compiler.

Macros

register_native_fn

Enums

KitlangError

Represents errors from different stages in the compilation process.

Functions

execute_source_string

Execute a given kitlang source code string with the MIR interpreter, using the provided native functions. The standard library is injected into the program with this function, if you do not want this for some reason, use execute_source_string_no_std instead. If time_execution is true, the different stages of execution will be timed and printed

execute_source_string_no_std

Execute a given kitlang source code string with the MIR interpreter, using the provided native functions. If time_execution is true, the different stages of execution will be timed and printed.

init_logging

Call this function to initialise logging, if the logging feature is enabled.

inject_standard_library

Inject the standard library code into a provided source code string.

parse_source_string_to_mir

Parse a given kitlang source code string into MIR representations. This function injects the standard library, if you do not want that, use parse_source_string_to_mir_no_std instead.

parse_source_string_to_mir_no_std

Parse a given kitlang source code string into MIR representations.

Type Aliases

KitSmallVec

A small vector type used throughout Kitlang for performance where a small vector is expected.

KitlangResult

Result type for Kitlang operations with error variants for different stages in the compilation process.