kitlang

Module ast

Defines the Abstract Syntax Tree (AST) data structures produced by the parser.

The AST represents the hierarchy of items in source code after parsing, things like Expressions being inside a Statement, which is inside a Function, which is inside a Module etc.

The module provides ASTRoot as the root container/node of the AST, IdentPath for module/type paths, and represents all possible expressions, statements, items, etc.

After parsing the source code is in a “complete” form syntactically, and fully describes what the program should do. However, it does not contain any semantic information, such as types of expressions, etc, and each expression that references a variable for example, still does so only by an identifier, and not by actually referencing the other node. This resolution is only really feasible once we lower the AST into HIR.

Structs

ASTRoot

This is the output of the Parser stage. Currently this only stores the AST from a single “root” module/project. In future this will store multiple modules for multiple file compilations.

Block

A declaration of a block, with a list of all statements inside the block.

Constant

A declaration of a constant, with the expression it should be evaluated to.

Enum

Enumeration Item in the AST, with a specified name Identifier, and all possible variants of the Enum.

EnumVariant

Describes one of the variants of an Enum, includes the name/Identifier of the variant and what kind of variant it is: Tuple, Struct or Unit.

Expression

An expression node in the AST.

FieldInitialisation

Describes the initialisation of a field within a StructInitialisation expression, contains the name/Identifier of the field to initialise, and an expression that will be the value of the field.

Function

A function Item declaration, holds information on the name of the function, all parameter info and the return type, with an optional body to the function, this is None if the function is just a declaration.

FunctionSig

The “signature” of a function, holds the return type and the parameters.

Ident

An Identifier. This is is defined as a unique type so it does not get confused as to what it describes.

IdentPath

A “path” consisting of Identifiers. You can think of this as a file system path like: “Users/Username/Downloads”, except this describes paths between different modules/files in our code, and instead of slashes we use the sequence of characters “::” as our seperator. So in our case this would be represented as "Users::Username::Downloads".

Impl

An Implementation block Item in the AST, contains for which Ty the impl block is for, as well as a list of all Item’s contained within.

Item

A “top-level” item in the AST. Functions, Constants, Structs, etc..

Local

A type of statement that declares a local variable, and optionally assigns it an initial value.

MethodCall

Describes a method call expression, contains the target of the method call, the Identifier specified which method to call, and the parameters to pass it.

Module

A Module Item in the AST, contains the name/Identifier of the module, as well as an optional body.

Parameter

An individual parameter to a function, includes the name Identifier, the Ty of the parameter, as well as if it is declared as mutable or not.

SourceSpan

Represents the span of bytes in the source string that the error originates from. This span is described as an exclusive span.

SpannedIdent

An identifier with an associated Span describing the range of bytes in the source code the identifier occupies.

SpannedIdentPath

Statement

A statement node within the AST.

Struct

A Struct Item in the AST, with an Identifier for it’s name, with all it’s associated fields.

StructField

A field of a struct, containing the name of the field and it’s associated Ty.

StructInitialisation

Describes an expression that initialises an instance of a Struct, contains the Ty of Struct to initialise, as well as the values to initialise all of the fields to.

UseImport

Describes a use Item in the AST.

Enums

BinaryOpKind

Kinds of “Binary” operations. A binary operation is an operation that acts on two values.

ExpressionAssociation

Describes the association of an expression. I.e. If the left-hand side or the right-hand side should be evaluated first.

ExpressionKind

Describes each possible kind of expressions.

ExpressionOrder

Describes the “order” or “precedence” of each type of expression. This is implemented as an enum with an ordering derive, so that the enum can be compared with itself to determine which of the expressions has higher precendence.

FunctionReturnTy

The return type of a function. If not specified this will default to FunctionReturnTy::Default which is the unit type (), otherwise this holds the specified return type.

ItemKind

Describes each possible kind of an Item in the AST.

Literal

Describes a kind of literal, with it’s associated parsed value.

LocalKind

Describes whether a local variable is just a declaration, or a declaration and an initial assignment.

ModuleKind

Describes the possible kinds of a “module”. This can either just be the declaration of a module, or a module Item with a body.

Mutability

Describes the mutability of a reference or variable.

StatementKind

Describes all possible kinds of statement in the AST.

Ty

Describes a Type. This structure is self-referential, I.e. an array will be described as a Ty::Array(Ty::Type(Identifier)).

UnaryOpKind

Kinds of “Unary” operations. A unary operation is an operation that acts on a single value.

VariantData

The kind of variant of an Enum.

Visibility

Describes whether an Item is able to be accessed/referenced from other modules/scopes.

Type Aliases

IdentPathSegment

IdentPathSegments