kitlang::ast

Enum ExpressionKind

pub enum ExpressionKind {

    Block(Box<Block>),
    Literal(Literal),
    BinaryOp(BinaryOpKind, Box<Expression>, Box<Expression>),
    UnaryOp(UnaryOpKind, Box<Expression>),
    If(Box<Expression>, Box<Block>, Option<Box<Expression>>),
    Loop(Box<Block>),
    For(SpannedIdent, Box<Expression>, Box<Block>),
    While(Box<Expression>, Box<Block>),
    Assign(Box<Expression>, Box<Expression>),
    Call(Box<Expression>, Vec<Box<Expression>>),
    MethodCall(Box<MethodCall>),
    Index(Box<Expression>, Box<Expression>),
    FieldAccess(Box<Expression>, Ident),
    StructInit(Box<StructInitialisation>),
    IdentPath(SpannedIdentPath),
    Continue,
    Break,
    Return(Option<Box<Expression>>),
    Cast(Box<Expression>, Ty),
    Range(Box<Expression>, Box<Expression>, bool),
    Tuple(Vec<Box<Expression>>),
    ArrayInit(Vec<Box<Expression>>),
    Reference(Box<Expression>, Mutability),

}

Describes each possible kind of expressions.

Variants

Block(Box<Block>)

A “block” of code.

Kit Example

<Expression>;
// Outer block..
let x = 20;
{
    // Inside of block..
    let y = 20;
}

Literal(Literal)

A Literal value, of either a number, boolean or string type.

Kit Example

20, "Hello!", true.

BinaryOp(BinaryOpKind, Box<Expression>, Box<Expression>)

Binary operation between two expressions.

Notes

The tuple values are: (BinaryOpKind, lhs, rhs).

Kit Example

20 + 70, 42 == 42, random_number() + 42.

UnaryOp(UnaryOpKind, Box<Expression>)

A unary operation on an expression.

Kit Example

!variable.

If(Box<Expression>, Box<Block>, Option<Box<Expression>>)

If statement, with a block or “body”, and an optional “else” case block.

Notes

The tuple values are: (condition_expression, if_true_block, else_block).

Kit Example

if random_number() == 20 {
    <Expression>;
}

Loop(Box<Block>)

An infinite loop, with a block or “body”.

For(SpannedIdent, Box<Expression>, Box<Block>)

For loop, with a block or “body” and an identifier for the loop variable.

Notes

The tuple values are: (variable_binding, iterable_expression, loop_block).

Kit Example

for i in 0..10 {
    println(i.to_string());
}

While(Box<Expression>, Box<Block>)

While loop, with a block or “body”, and an optional “else” case block.

Notes

The tuple values are: (condition_expression, loop_block).

Kit Example

while i > 0 {
    i = i - 1
}

Assign(Box<Expression>, Box<Expression>)

Assignment of an already declared value.

Notes

The tuple values are: (variable_to_be_assigned_to, new_value)

Kit Example

x = x + 1

Call(Box<Expression>, Vec<Box<Expression>>)

Calling a free function, or using the call operator of an object.

Notes

The tuple values are: (call_target, parameters)

Kit Example

random_number(0, 20)

MethodCall(Box<MethodCall>)

Call a method function of an object, or the ‘dot’ operator.

Kit Example

vec2.to_string()

Index(Box<Expression>, Box<Expression>)

Index into an array or an object that implements the indexing operator.

Notes

The tuple values are: (object_to_be_indexed, index)

Kit Example

list_of_names[2]

FieldAccess(Box<Expression>, Ident)

Access a data field on an object.

Notes

The tuple values are: (object_to_access, field_to_access)

Kit Examples

vec2.x

StructInit(Box<StructInitialisation>)

Initialise a struct with specified field values.

Kit Example

Vec2 {
    x: 10,
    y: 20
}

IdentPath(SpannedIdentPath)

An Identifier or Path.

Kit Example

Module::Path::To::Function, x.

Continue

Skip the rest of the loop code and proceed to the next iteration early.

Break

Break out of the contained loop.

Return(Option<Box<Expression>>)

Return from the function, with an optional expression to use as the return value.

Cast(Box<Expression>, Ty)

Cast an expression to another type.

Range(Box<Expression>, Box<Expression>, bool)

A range expression.

Tuple(Vec<Box<Expression>>)

A tuple expression with multiple elements.

Kit Example

(1, 2, 3), (x, y, name).

ArrayInit(Vec<Box<Expression>>)

A tuple expression with multiple elements.

Kit Example

(1, 2, 3), (x, y, name).

Reference(Box<Expression>, Mutability)

A reference to a value.

Kit Example

let x = 20;
let y = &x; // Reference to x
let z = &mut x; // Mutable reference to x

Implementations

impl ExpressionKind

pub const fn can_be_non_semi(&self) -> bool

Returns true if the expression is allowed to omit the trailing semi-colon without being the last statement in a block.

pub const fn get_order(&self) -> ExpressionOrder

Get the “precendence” of the expression.

pub const fn get_association(&self) -> ExpressionAssociation

Get the “association” of the expression.