kitlang::compiler

Struct Compiler

pub struct Compiler {
    pub context: CompilerContext,
    pub profile_stages: bool,
}

The main compiler structure, holds the compiler context and compilation state and logic.

Fields
context: CompilerContextprofile_stages: bool
Implementations

impl Compiler

pub fn new(source: impl AsRef<str>) -> Self

Create a new compiler with the given source code.

pub fn new_with_profiler(source: impl AsRef<str>) -> Self

Create a new compiler with the given source code, with profiling enabled.

impl Compiler

pub fn profile_stages(self, should_profile: bool) -> Self

Profile the different compilation stages.

pub fn no_stdlib(self, no_stdlib: bool) -> Self

Set whether to compile without the standard library.

impl Compiler

pub fn tokenise(&self) -> impl Iterator<Item = Token>

Tokenise the source code into a stream of tokens. This function does not skip whitespace or comments. If this behaviour is desired, use tokenise_stripped instead.

Note

This will collect al tokens into a vector before returning an iterator, so may use more memory for large source files.

Returns

An iterator over the tokens in the source code.

pub fn tokenise_stripped(&self) -> impl Iterator<Item = Token>

Tokenise the source code into a stream of tokens. This function skips whitespace and comments. If this behaviour is not desired, use tokenise instead.

Note

This will collect al tokens into a vector before returning an iterator, so may use more memory for large source files.

Returns

An iterator over the tokens in the source code.

pub fn parse_ast(&self) -> KitlangResult<ASTRoot>

Parse up to the AST stage from the source code provided in the compiler context.

Errors

This function will return an error if any part of parsing the token stream fails. The returned error will contain information about where in the source code the error occurred, and a message about the nature of the error.

Returns

Ok(ASTRoot) - The parsed AST root node if parsing was successful.

Err(KitlangError) - An error indicating what went wrong during parsing

pub fn compile_to_hir(&mut self) -> KitlangResult<HLIR>

Compile the provided source code up to the HIR stage.

Note

This function will do all the required resolving and type checking to produce a valid HIR.

Errors

This function will return an error if any part of the AST cannot be lowered properly. The returned error will contain a diagnostic message indicating the nature and location of any failures, as well as which stage of lowering it occurred in.

Returns

Ok(HLIR) - The lowered HIR if successful.

Err(KitlangError) - An error indicating what went wrong during lowering

pub fn compile_to_mir(&mut self) -> KitlangResult<MIR>

Compile the provided source code up to the MIR stage. This function will parse to AST, then compile to HIR, then lower the HIR to MIR.

Errors

This function will return an error if any part of the HIR cannot be lowered properly. The returned error will contain a diagnostic message indicating the nature and location of any failures.

Returns

Ok(MIR) - The lowered MIR if successful.

Err(KitlangError) - An error indicating what went wrong during any lowering or parsing stage.