kitlang::compiler

Struct CompilerContext

pub struct CompilerContext {
    source: String,
    pub meta: ProgramMetaData,
    pub no_stdlib: bool,
}

A context structure for the compiler, holds state relevant to the compilation process.

Fields
source: String

The source code being compiled.

meta: ProgramMetaData

Metadata about the program being compiled.

no_stdlib: bool

Whether to compile without the standard library.

Implementations

impl CompilerContext

pub fn new(source: String) -> Self

Create a new compiler context with the given source code and metadata.

impl CompilerContext

pub fn meta(&self) -> &ProgramMetaData

Get a reference to the program’s meta data.

pub fn meta_mut(&mut self) -> &mut ProgramMetaData

Get a mutable reference to the program’s namespace.

pub fn namespace(&self) -> &Namespace

Get a reference to the program’s namespace.

pub fn namespace_mut(&mut self) -> &mut Namespace

Get a mutable reference to the program’s namespace.

pub fn type_registry(&self) -> &TypeRegistry

Get a reference to the program’s type registry.

pub fn type_registry_mut(&mut self) -> &mut TypeRegistry

Get a mutable reference to the program’s type registry.

pub fn type_map(&self) -> &TypeMap

Get a reference to the program’s type registry.

pub fn type_map_mut(&mut self) -> &mut TypeMap

Get a mutable reference to the program’s type registry.

impl CompilerContext

const CORE_LIB: &str = "// @lang\nimpl i8 { \n pub fn to_string(self) -> string { i64_to_string(self as i64) }\n pub fn sqrt(self) -> i8 { i64_sqrt(self as i64) as i8 }\n pub fn abs(self) -> i8 { i64_abs(self as i64) as i8 }\n}\n// @lang\nimpl i16 { \n pub fn to_string(self) -> string { i64_to_string(self as i64) }\n pub fn sqrt(self) -> i16 { i64_sqrt(self as i64) as i16 }\n pub fn abs(self) -> i16 { i64_abs(self as i64) as i16 }\n}\n// @lang\nimpl i32 { \n pub fn to_string(self) -> string { i64_to_string(self as i64) }\n pub fn sqrt(self) -> i32 { i64_sqrt(self as i64) as i32 }\n pub fn abs(self) -> i32 { i64_abs(self as i64) as i32 }\n}\n// @lang\nimpl i64 { \n pub fn to_string(self) -> string { i64_to_string(self) }\n pub fn sqrt(self) -> i64 { i64_sqrt(self) }\n pub fn abs(self) -> i64 { i64_abs(self) }\n}\n// @lang\nimpl i128 { \n pub fn to_string(self) -> string { i64_to_string(self as i64) }\n pub fn sqrt(self) -> i128 { i64_sqrt(self as i64) as i128 }\n pub fn abs(self) -> i128 { i64_abs(self as i64) as i128 }\n}\n\n// @lang\nimpl u8 { \n pub fn to_string(self) -> string { u64_to_string(self as u64) }\n pub fn sqrt(self) -> u8 { u64_sqrt(self as u64) as u8 }\n}\n// @lang\nimpl u16 { \n pub fn to_string(self) -> string { u64_to_string(self as u64) }\n pub fn sqrt(self) -> u16 { u64_sqrt(self as u64) as u16 }\n}\n// @lang\nimpl u32 { \n pub fn to_string(self) -> string { u64_to_string(self as u64) }\n pub fn sqrt(self) -> u32 { u64_sqrt(self as u64) as u32 }\n}\n// @lang\nimpl u64 { \n pub fn to_string(self) -> string { u64_to_string(self) }\n pub fn sqrt(self) -> u64 { u64_sqrt(self) }\n}\n// @lang\nimpl u128 { \n pub fn to_string(self) -> string { u64_to_string(self as u64) }\n pub fn sqrt(self) -> u128 { u64_sqrt(self as u64) as u128 }\n}\n\n// @lang\nimpl f32 {\n pub fn to_string(self) -> string { f64_to_string(self as f64) }\n pub fn sqrt(self) -> f32 { f64_sqrt(self as f64) as f32 }\n pub fn abs(self) -> f32 { f64_abs(self as f64) as f32 }\n}\n// @lang\nimpl f64 {\n pub fn to_string(self) -> string { f64_to_string(self) }\n pub fn sqrt(self) -> f64 { f64_sqrt(self) }\n pub fn abs(self) -> f64 { f64_abs(self) }\n}\n\n// @lang\nimpl bool {\n pub fn to_string(self) -> string { bool_to_string(self) }\n}\n\n// @lang\nimpl string {\n pub fn to_string(self) -> string { self }\n}\n\n// @lang\nimpl () {\n pub fn to_string(self) -> string { \"()\" }\n}"

const INTRINSICS_LIB: &str = "native global fn to_lower(s: string) -> string;\nnative global fn is_empty(s: string) -> bool;\n\nnative global fn print(s: string);\nnative global fn println(s: string);\n\nnative global fn i64_to_string(i: i64) -> string;\nnative global fn string_to_i64(s: string) -> i64;\nnative global fn u64_to_string(u: u64) -> string;\nnative global fn string_to_u64(s: string) -> u64;\nnative global fn f64_to_string(f: f64) -> string;\nnative global fn string_to_f64(s: string) -> f64;\nnative global fn bool_to_string(f: bool) -> string;\nnative global fn string_to_bool(s: string) -> bool;\n\nnative global fn i64_sqrt(i: i64) -> i64;\nnative global fn u64_sqrt(u: u64) -> u64;\nnative global fn f64_sqrt(f: f64) -> f64;\n\nnative global fn i64_abs(i: i64) -> i64;\nnative global fn f64_abs(f: f64) -> f64;"

pub fn source(&self) -> String

Get the full source code, with standard library injected if applicable.

pub fn raw_source(&self) -> &str

Get the raw source code, without standard library injected, regardless of settings.