kitlang::intermediate

Module hir

The HIR module provides the primary IR structure for the compiler. It transforms the AST into a flatter representation optimized for analysis passes like resolution and type checking.

The HLIR organizes nodes into “owning nodes” (top-level items like functions, structs, impls) and their owned nodes (expressions, statements), each addressable via HirId. This structure enables efficient lookups and storage of side-channel data like type information without deep tree traversals, and having to store a way of navigating this tree structure.

This module defines core ID types (HirId, OwnerDefId, LocalDefId), the HLIR container, program metadata (ProgramMetaData), and utilities for lowering, traversing, and manipulating the HIR.

Re-exports

pub use crate::intermediate::hir::errors::LoweringError;

pub use crate::intermediate::hir::errors::LoweringErrorKind;

Modules

errors

lowerer

HIR Lowering

nodes

visitor

Structs

DefId

Definition ID that is relative to a specified module.

HLIR

High Level Intermediate Representation of the program.

HirId

Definition ID that describes which ID into the modules "path_table" owns the node, and which index into the owners body the node refers to.

LocalDefId

Definition ID that is only valid when relative to the “owner”.

OwnerDefId

Definition ID that describes the index into the global “module” nodes of an owning HIR node.

ProgramMetaData

Contains metadata about the program stored in HIR, such as type information, namespace etc.

Functions

get_span_by_id 🔒

Get the source span for a given HIR ID.

lower_ast_to_hir

Lower the output of the parser stage to HIR.

parse_ast_to_hir_processed

Lower the output of the parser state to HIR.

parse_ast_to_hir_processed_with

Lower the output of the parser state to HIR, using a provided metadata structure.