kitlang::intermediate::hir::hlir

Trait HLIRExt

pub trait HLIRExt<'a, 'b> {

    // Required methods
    fn owner_node_count(&self) -> u32;
    fn owning_node(&'a self, id: OwnerDefId) -> Option<&'b OwningNode>;
    fn owning_node_mut(
        &'a mut self,
        id: OwnerDefId,
    ) -> Option<&'b mut OwningNode>;
    fn owning_node_unchecked(&'a self, id: OwnerDefId) -> &'b OwningNode;
    fn owning_node_mut_unchecked(
        &'a mut self,
        id: OwnerDefId,
    ) -> &'b mut OwningNode;

    // Provided methods
    fn next_owner_id(&self) -> OwnerDefId { ... }
    fn root_module(&'a self) -> Option<&'b OwningNode> { ... }
    fn root_module_mut(&'a mut self) -> Option<&'b mut OwningNode> { ... }
    fn owner_id_iter(&self) -> impl Iterator<Item = OwnerDefId> { ... }
    fn get_hir_node(&'a self, hir_id: HirId) -> Option<&'b HirNode> { ... }
    fn get_hir_node_mut(&'a mut self, hir_id: HirId) -> Option<&'b mut HirNode> { ... }
    fn span_by_owner_id(&'a self, id: OwnerDefId) -> Option<SourceSpan> { ... }
    fn span_by_hir_id(&'a self, id: HirId) -> Option<SourceSpan> { ... }
    fn binding_by_id(&'a self, id: HirId) -> Option<&'b VarBinding> { ... }
    fn owning_node_item(&'a self, owner_id: OwnerDefId) -> Option<&'b Item> { ... }
    fn owning_node_item_mut(
        &'a mut self,
        owner_id: OwnerDefId,
    ) -> Option<&'b mut Item> { ... }
    fn owning_node_item_as<F>(&'a self, owner_id: OwnerDefId) -> Option<&'b F>
       where Option<&'b F>: From<&'b OwningNode> { ... }
    fn owning_node_item_mut_as<F>(
        &'a mut self,
        owner_id: OwnerDefId,
    ) -> Option<&'b mut F>
       where Option<&'b mut F>: From<&'b mut OwningNode> { ... }
    fn get_hir_node_unchecked(&'a self, hir_id: HirId) -> &'b HirNode { ... }
    fn get_hir_node_mut_unchecked(
        &'a mut self,
        hir_id: HirId,
    ) -> &'b mut HirNode { ... }
    fn get_hir_node_as<F>(&'a self, id: HirId) -> Option<&'b F>
       where Option<&'b F>: From<&'b HirNode> { ... }
    fn get_hir_node_mut_as<F>(&'a mut self, id: HirId) -> Option<&'b mut F>
       where Option<&'b mut F>: From<&'b mut HirNode> { ... }

}

Required Methods

fn owner_node_count(&self) -> u32

Get the total count of owner nodes in the HIR.

fn owning_node(&'a self, id: OwnerDefId) -> Option<&'b OwningNode>

Get a reference to the owning node with a specified OwnerDefId.

fn owning_node_mut(&'a mut self, id: OwnerDefId) -> Option<&'b mut OwningNode>

Get a mutable reference to the owning node with a specified OwnerDefId.

fn owning_node_unchecked(&'a self, id: OwnerDefId) -> &'b OwningNode

Get a reference to the owning node without checking for existence.

Panics

Panics if the node does not exist.

fn owning_node_mut_unchecked(&'a mut self, id: OwnerDefId) -> &'b mut OwningNode

Get a mutable reference to the owning node without checking for existence.

Panics

Panics if the node does not exist.

Provided Methods

fn next_owner_id(&self) -> OwnerDefId

Get the next available owner ID for inserting a new owning node.

fn root_module(&'a self) -> Option<&'b OwningNode>

Get a reference to the root module owning node.

fn root_module_mut(&'a mut self) -> Option<&'b mut OwningNode>

Get a mutable reference to the root module owning node.

fn owner_id_iter(&self) -> impl Iterator<Item = OwnerDefId>

Get an iterator over all owner IDs in the HIR.

fn get_hir_node(&'a self, hir_id: HirId) -> Option<&'b HirNode>

Get a reference to the HirNode with a specified HirId.

fn get_hir_node_mut(&'a mut self, hir_id: HirId) -> Option<&'b mut HirNode>

Get a mutable reference to the HirNode with a specified HirId.

fn span_by_owner_id(&'a self, id: OwnerDefId) -> Option<SourceSpan>

Get the source span for a given owning node ID, if it exists.

fn span_by_hir_id(&'a self, id: HirId) -> Option<SourceSpan>

Get the source span for a given HirId, if it exists.

fn binding_by_id(&'a self, id: HirId) -> Option<&'b VarBinding>

Get a reference to a variable binding by its HirId, if the node exists and is a binding.

fn owning_node_item(&'a self, owner_id: OwnerDefId) -> Option<&'b Item>

Get a reference to the owning node’s item with a specified OwnerDefId.

fn owning_node_item_mut( &'a mut self, owner_id: OwnerDefId, ) -> Option<&'b mut Item>

Get a mutable reference to the owning node’s item with a specified OwnerDefId.

fn owning_node_item_as<F>(&'a self, owner_id: OwnerDefId) -> Option<&'b F>
where Option<&'b F>: From<&'b OwningNode>,

Get a reference to the owning node’s item, cast to type F, (if possible) with a specified OwnerDefId.

fn owning_node_item_mut_as<F>( &'a mut self, owner_id: OwnerDefId, ) -> Option<&'b mut F>
where Option<&'b mut F>: From<&'b mut OwningNode>,

Get a mutable reference to the owning node’s item, cast to type F, (if possible) with a specified OwnerDefId.

fn get_hir_node_unchecked(&'a self, hir_id: HirId) -> &'b HirNode

Get a reference to the HirNode with a specified HirId without checking for existence.

Panics

Panics if the node does not exist.

fn get_hir_node_mut_unchecked(&'a mut self, hir_id: HirId) -> &'b mut HirNode

Get a mutable reference to the HirNode with a specified HirId without checking for existence.

Panics

Panics if the node does not exist.

fn get_hir_node_as<F>(&'a self, id: HirId) -> Option<&'b F>
where Option<&'b F>: From<&'b HirNode>,

Get a reference to the HirNode with a specified HirId, casted to type F.

fn get_hir_node_mut_as<F>(&'a mut self, id: HirId) -> Option<&'b mut F>
where Option<&'b mut F>: From<&'b mut HirNode>,

Get a mutable reference to the HirNode with a specified HirId, casted to type F.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.