kitlang::intermediate::hir

Module lowerer

HIR Lowering

The purpose of this module is to lower the output of the previous stage (Parser) which is the AST, into a High-Level Intermediate Representation (HIR).

The reason for having a separate HIR phase is to simplify and normalize the structure of the code, making it easier to perform subsequent analyses such as name resolution and type checking.

The main effect of this lowering phase is essentially to “flatten” the AST structure into something that is more “linear” in a sense, where each node is not nested inside another node, but instead each node is assigned an ID, which is essentially an index, which makes it much easier to reference nodes from other nodes, and to have “side” data such as type information associated with a node, without the node itself storing it.

Overview

After lowering is performed, no subsequent steps are yet performed, I.e. everything is still unresolved, but the structure is now more suitable for analysis.

The main steps following lowering to HIR that are performed are:

Name Resolution

Type Checking

Owning Nodes vs HIR Nodes

Owning Nodes: Important items like functions, structs, modules, and impl blocks that can contain other nodes, and are said to “own” those nodes that are inside of them, each owning node receives a unique OwnerDefId. An OwnerDefId is essentially an index into the HLIR’s owning node storage.

HIR Nodes: Expressions, statements, parameters, etc. that exist within an owning node. These receive HirIds scoped to their owner, where a HirId is essentially both an OwnerDefId and an index within that owning node’s HIR node storage.

Structs

HLIRLowerer 🔒

This struct is how we store and maintain the HLIR while we lower from AST to HIR. Contains all the logic for doing the lowering.

Functions

lower_ast_to_hir

Lowers an Abstract Syntax Tree to High-Level Intermediate Representation.