kitlang::intermediate::hir

Module nodes

Macros

impl_hir_node_from 🔒

impl_item_kind_shorthand 🔒

Structs

BindingModifiers

Modifiers for a variable binding, such as mutability.

Block

A block of statements in HIR. Holds a list of HirIds that make up the block and whether it is the root block of an item or not.

Constant

A declaration of a constant, with the expression it should be evaluated to.

Enum

An enum in HIR, containing its identifier and other metadata.

Expr

An expression in HIR, containing its kind and source span.

Function

A function in HIR, containing its signature, body, and other metadata.

FunctionBody

The body of a Function, containing the parameter IDs and the block ID.

FunctionSig

The “signature” of a function, holds the return type and the parameter types.

Impl

An impl block in HIR, containing the type being implemented on and the items within the block.

Item

An item in HIR, containing its owner definition ID and kind.

LetStatement

A let statement in HIR, containing the identifier, mutability, type, and optional initial value.

Module

A module in HIR, with a list of OwnerDefIds of the items it contains.

OwningNode

Owning HIR node, which contains other HIR nodes and manages their scope. Has a kind, which describes what type of owning node it is (e.g. an Item or an Item within an impl block).

Parameter

An individual parameter to a function.

Statement

A statement in HIR, containing its kind and source span.

Struct

A struct in HIR, containing its identifier, fields, and other metadata.

StructField

A field of a struct, containing the name of the field and it’s associated Type.

StructFieldInit

A field initialisation within a struct initialisation expression.

StructInitialisation

A struct initialisation expression in HIR, containing the type path and field initialisations.

UsePath

A use path in HIR, containing the path being imported and other metadata.

VarBinding

A variable binding in HIR, containing its HirId and kind.

Enums

BindingKind

The different kinds of variable bindings in HIR. I.e. A single identifier such as let x = ... or a tuple binding such as let (x, y) = .... In this system, the Ident variant is used as the atomic unit of variable bindings. Tuple bindings are represented as a collection of Ident bindings, referenced by a HirId.

ExprKind

The different kinds of expressions in HIR.

HirNode

A node within an owning HIR node. Represents things such as statements, expressions, parameters, etc.

ItemKind

The different kinds of items in HIR. Differentiates between modules, functions, structs, constants, etc.

ModuleIdent

Identifier for a module, either raw or spanned.

ModuleSpan

Describes the span of a module, either declaration or implementation. Used to differentiate between the two in error messages.

OwningNodeKind

Describes a kind of node that “controls” a scope and owns it’s contents.

RefPath

A reference path in HIR, which can be either unresolved (AST path) or resolved (with a ResolvedID).

ResolvedID

The resolved ID of a reference path in HIR.

StatementKind

The different kinds of statements in HIR.

Type

The type of a value in HIR, either unresolved (AST type) or resolved (Kit type).