kitlang::intermediate::hir::nodes

Struct OwningNode

pub struct OwningNode {
    pub kind: OwningNodeKind,
    pub nodes: Vec<HirNode>,
}

Owning HIR node, which contains other HIR nodes and manages their scope. Has a kind, which describes what type of owning node it is (e.g. an Item or an Item within an impl block).

Fields
kind: OwningNodeKindnodes: Vec<HirNode>

The HIR nodes owned by this owning node.

Implementations

impl OwningNode

pub const fn new(kind: OwningNodeKind) -> Self

Create a new OwningNode with the specified kind.

pub const fn from_item(item: Item, impl_item: bool) -> Self

Create a new OwningNode from an Item, specifying whether it is an item within an impl block.

pub const fn set_owner_id(&mut self, owner_id: OwnerDefId)

Set the owner definition ID of the owning node.

pub const fn owner_id(&self) -> OwnerDefId

Get the owner definition ID of the owning node.

pub fn ident(&self) -> Option<String>

Get the identifier of the item in the owning node, if it has one.

pub const fn item(&self) -> &Item

Get a reference to the item in the owning node.

pub const fn item_mut(&mut self) -> &mut Item

Get a mutable reference to the item in the owning node.

pub const fn span(&self) -> SourceSpan

Get the source span of the item in the owning node, if it has one.

impl OwningNode

pub const fn local_count(&self) -> u32

Get the number of local HIR nodes owned by this owning node.

pub const fn next_local_id(&self) -> LocalDefId

Get the next available local definition ID for a new HIR node.

pub const fn next_hir_id(&self) -> HirId

Get the next available HIR ID for a new HIR node. Combines the owning node’s owner ID with the next local definition ID.

pub fn insert_hir_node(&mut self, hir_node: HirNode) -> LocalDefId

Insert a new HIR node into the owning node, returning its assigned local definition ID.

pub fn get_hir_node(&self, id: LocalDefId) -> Option<&HirNode>

Get a reference to a HIR node by its local definition ID.

pub fn get_hir_node_mut(&mut self, id: LocalDefId) -> Option<&mut HirNode>

Get a mutable reference to a HIR node by its local definition ID.

impl OwningNode

pub const fn hir_module_mut(&mut self) -> Option<&mut Module>

pub const fn hir_module_ref(&self) -> Option<&Module>

impl OwningNode

pub const fn hir_function_mut(&mut self) -> Option<&mut Function>

pub const fn hir_function_ref(&self) -> Option<&Function>

impl OwningNode

pub const fn hir_enum_mut(&mut self) -> Option<&mut Enum>

pub const fn hir_enum_ref(&self) -> Option<&Enum>

impl OwningNode

pub const fn hir_struct_mut(&mut self) -> Option<&mut Struct>

pub const fn hir_struct_ref(&self) -> Option<&Struct>

impl OwningNode

pub const fn hir_const_mut(&mut self) -> Option<&mut Constant>

pub const fn hir_const_ref(&self) -> Option<&Constant>

impl OwningNode

pub const fn hir_impl_mut(&mut self) -> Option<&mut Impl>

pub const fn hir_impl_ref(&self) -> Option<&Impl>

impl OwningNode

pub const fn hir_use_mut(&mut self) -> Option<&mut UsePath>

pub const fn hir_use_ref(&self) -> Option<&UsePath>