kitlang::intermediate::hir::visitor

Trait HLIRVisitor

pub trait HLIRVisitor {

    // Provided methods
    fn super_root(&mut self, hlir: &HLIR) { ... }
    fn super_module(&mut self, module: &Module, hlir: &HLIR) { ... }
    fn super_item(&mut self, item: &Item, hlir: &HLIR) { ... }
    fn super_impl_item(&mut self, item: &Item, hlir: &HLIR) { ... }
    fn super_use(&mut self, _use_info: &UsePath, _hlir: &HLIR) { ... }
    fn super_impl(&mut self, impl_info: &Impl, hlir: &HLIR) { ... }
    fn super_function(&mut self, function: &Function, hlir: &HLIR) { ... }
    fn super_function_param(&mut self, _parameter: &Parameter, _hlir: &HLIR) { ... }
    fn super_enum(&mut self, _enumeration: &Enum, _hlir: &HLIR) { ... }
    fn super_struct(&mut self, _structure: &Struct, _hlir: &HLIR) { ... }
    fn super_block(&mut self, block: &Block, hlir: &HLIR) { ... }
    fn super_statement(
        &mut self,
        statement: &Statement,
        _parent_block: &Block,
        hlir: &HLIR,
    ) { ... }
    fn super_let_statement(
        &mut self,
        _id: HirId,
        let_statement: &LetStatement,
        hlir: &HLIR,
    ) { ... }
    fn super_constant(&mut self, constant: &Constant, hlir: &HLIR) { ... }
    fn super_expr(&mut self, expr: &Expr, hlir: &HLIR) { ... }
    fn super_path(&mut self, _hir_id: HirId, _path: &RefPath, _hlir: &HLIR) { ... }
    fn visit_expr_by_id(&mut self, expr_id: HirId, hlir: &HLIR) { ... }
    fn visit_block_by_id(&mut self, block_id: HirId, hlir: &HLIR) { ... }
    fn visit_path(&mut self, hir_id: HirId, path: &RefPath, hlir: &HLIR) { ... }
    fn visit_expr(&mut self, expr: &Expr, hlir: &HLIR) { ... }
    fn visit_constant(&mut self, constant: &Constant, hlir: &HLIR) { ... }
    fn visit_let_statement(
        &mut self,
        id: HirId,
        let_statement: &LetStatement,
        hlir: &HLIR,
    ) { ... }
    fn visit_statement(
        &mut self,
        statement: &Statement,
        parent_block: &Block,
        hlir: &HLIR,
    ) { ... }
    fn visit_block(&mut self, block: &Block, hlir: &HLIR) { ... }
    fn visit_struct(&mut self, structure: &Struct, hlir: &HLIR) { ... }
    fn visit_enum(&mut self, enumeration: &Enum, hlir: &HLIR) { ... }
    fn visit_function_param(&mut self, parameter: &Parameter, hlir: &HLIR) { ... }
    fn visit_function(&mut self, function: &Function, hlir: &HLIR) { ... }
    fn visit_impl(&mut self, impl_info: &Impl, hlir: &HLIR) { ... }
    fn visit_use(&mut self, use_info: &UsePath, hlir: &HLIR) { ... }
    fn visit_item(&mut self, item: &Item, hlir: &HLIR) { ... }
    fn visit_impl_item(&mut self, item: &Item, hlir: &HLIR) { ... }
    fn visit_module(&mut self, module: &Module, hlir: &HLIR) { ... }
    fn visit_root(&mut self, hlir: &HLIR) { ... }

}

Provides an interface from traversing the HLIR tree.

Functions named super_ are default implementations that traverse the tree further, and are not meant to be overriden. They should be called during the corresponding visit_ function implementations to keep traversing the tree further, you can also omit this under conditions if you do not want traversal to continue.

The super_ functions will call the appropriate visit_ functions for child nodes, which should then call the super_ functions to continue traversal, if desired.

By default all visit_ functions call their corresponding super_ functions, so only the kinds of nodes you want to handle need to be overridden/implemented.

Implementations should be provided by overriding the visit_ functions for the specified types you want.

To start traversal, the function visit_root should be called on this trait with a reference to the HLIR.

Provided Methods

fn super_root(&mut self, hlir: &HLIR)

Default implementation for visiting the root of the HLIR. This visits the root module if it exists.

fn super_module(&mut self, module: &Module, hlir: &HLIR)

Default implementation for visiting a module. This visits all items within the module.

fn super_item(&mut self, item: &Item, hlir: &HLIR)

Default implementation for visiting an item. This dispatches to the appropriate visit function based on the item kind.

fn super_impl_item(&mut self, item: &Item, hlir: &HLIR)

Default implementation for visiting an item within an impl block. This simply calls the super_item function.

fn super_use(&mut self, _use_info: &UsePath, _hlir: &HLIR)

Default implementation for visiting a use path. This does nothing by default.

fn super_impl(&mut self, impl_info: &Impl, hlir: &HLIR)

Default implementation for visiting an impl block. This visits all items within the impl block.

fn super_function(&mut self, function: &Function, hlir: &HLIR)

Default implementation for visiting a function. This visits the function parameters and the function body block.

fn super_function_param(&mut self, _parameter: &Parameter, _hlir: &HLIR)

Default implementation for visiting a function parameter. This does nothing by default.

fn super_enum(&mut self, _enumeration: &Enum, _hlir: &HLIR)

Default implementation for visiting an enum. This does nothing by default.

fn super_struct(&mut self, _structure: &Struct, _hlir: &HLIR)

Default implementation for visiting a struct. This does nothing by default.

fn super_block(&mut self, block: &Block, hlir: &HLIR)

Default implementation for visiting a block. This visits all statements within the block.

fn super_statement( &mut self, statement: &Statement, _parent_block: &Block, hlir: &HLIR, )

Default implementation for visiting a statement. This dispatches based on the statement kind, either visiting a let statement, and item or an expression.

fn super_let_statement( &mut self, _id: HirId, let_statement: &LetStatement, hlir: &HLIR, )

Default implementation for visiting a let statement. This visits the initial value expression if it exists.

fn super_constant(&mut self, constant: &Constant, hlir: &HLIR)

Default implementation for visiting a constant. This visits the expression of the constant.

fn super_expr(&mut self, expr: &Expr, hlir: &HLIR)

Default implementation for visiting an expression. This dispatches based on the expression kind, visiting child expressions/blocks as necessary.

fn super_path(&mut self, _hir_id: HirId, _path: &RefPath, _hlir: &HLIR)

Default implementation for visiting a path. This does nothing by default.

fn visit_expr_by_id(&mut self, expr_id: HirId, hlir: &HLIR)

Helper function to call visit_expr by using a given HirId.

fn visit_block_by_id(&mut self, block_id: HirId, hlir: &HLIR)

Helper function to call visit_block by using a given HirId.

fn visit_path(&mut self, hir_id: HirId, path: &RefPath, hlir: &HLIR)

Called for every path in the HLIR.

fn visit_expr(&mut self, expr: &Expr, hlir: &HLIR)

Called for every expression in the HLIR.

fn visit_constant(&mut self, constant: &Constant, hlir: &HLIR)

Called for every constant in the HLIR.

fn visit_let_statement( &mut self, id: HirId, let_statement: &LetStatement, hlir: &HLIR, )

Called for every let statement in the HLIR.

fn visit_statement( &mut self, statement: &Statement, parent_block: &Block, hlir: &HLIR, )

Called for every statement in the HLIR.

fn visit_block(&mut self, block: &Block, hlir: &HLIR)

Called for every block in the HLIR.

fn visit_struct(&mut self, structure: &Struct, hlir: &HLIR)

Called for every struct in the HLIR.

fn visit_enum(&mut self, enumeration: &Enum, hlir: &HLIR)

Called for every enum in the HLIR.

fn visit_function_param(&mut self, parameter: &Parameter, hlir: &HLIR)

Called for every function parameter in the HLIR.

fn visit_function(&mut self, function: &Function, hlir: &HLIR)

Called for every function in the HLIR.

fn visit_impl(&mut self, impl_info: &Impl, hlir: &HLIR)

Called for every impl block in the HLIR.

fn visit_use(&mut self, use_info: &UsePath, hlir: &HLIR)

Called for every use path in the HLIR.

fn visit_item(&mut self, item: &Item, hlir: &HLIR)

Called for every item in the HLIR.

fn visit_impl_item(&mut self, item: &Item, hlir: &HLIR)

Called for every item within an impl block in the HLIR.

fn visit_module(&mut self, module: &Module, hlir: &HLIR)

Called for every module in the HLIR.

fn visit_root(&mut self, hlir: &HLIR)

Entry point for visiting the HLIR.