kitlang::intermediate::hir::visitor

Trait HLIRVisitorMut

pub trait HLIRVisitorMut<'a> {

    // Provided methods
    fn super_root_mut(&mut self, hlir: &mut HLIRDisjointMut<'a>) { ... }
    fn super_module_mut(
        &mut self,
        module: &mut Module,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn super_item_mut(
        &mut self,
        item: &mut Item,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn super_impl_item_mut(
        &mut self,
        item: &mut Item,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn super_use_mut(
        &mut self,
        _use_info: &mut UsePath,
        _hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn super_impl_mut(
        &mut self,
        impl_info: &mut Impl,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn super_function_param_mut(
        &mut self,
        _parameter: &mut Parameter,
        _hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn super_function_mut(
        &mut self,
        function: &mut Function,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn super_enum_mut(
        &mut self,
        _enumeration: &mut Enum,
        _hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn super_struct_mut(
        &mut self,
        _structure: &Struct,
        _hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn super_block_mut(&mut self, block: &Block, hlir: &mut HLIRDisjointMut<'a>) { ... }
    fn super_statement_mut(
        &mut self,
        statement: &mut Statement,
        _parent_block: &Block,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn super_let_statement_mut(
        &mut self,
        _id: HirId,
        let_statement: &mut LetStatement,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn super_constant_mut(
        &mut self,
        constant: &mut Constant,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn super_expr_mut(
        &mut self,
        expr: &mut Expr,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn super_path_mut(
        &mut self,
        _id: HirId,
        _path: &mut RefPath,
        _hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_expr_by_id_mut(
        &mut self,
        expr_id: HirId,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_block_by_id_mut(
        &mut self,
        block_id: HirId,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_path_mut(
        &mut self,
        hir_id: HirId,
        path: &mut RefPath,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_expr_mut(
        &mut self,
        expr: &mut Expr,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_constant_mut(
        &mut self,
        constant: &mut Constant,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_let_statement_mut(
        &mut self,
        id: HirId,
        let_statement: &mut LetStatement,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_statement_mut(
        &mut self,
        statement: &mut Statement,
        parent_block: &Block,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_block_mut(
        &mut self,
        block: &mut Block,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_struct_mut(
        &mut self,
        structure: &mut Struct,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_enum_mut(
        &mut self,
        enumeration: &mut Enum,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_function_mut(
        &mut self,
        function: &mut Function,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_function_param_mut(
        &mut self,
        parameter: &mut Parameter,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_impl_mut(
        &mut self,
        impl_info: &mut Impl,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_use_mut(
        &mut self,
        use_info: &mut UsePath,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_item_mut(
        &mut self,
        item: &mut Item,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_impl_item_mut(
        &mut self,
        item: &mut Item,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_module_mut(
        &mut self,
        module: &mut Module,
        hlir: &mut HLIRDisjointMut<'a>,
    ) { ... }
    fn visit_root_mut(&mut self, hlir: &mut HLIRDisjointMut<'a>) { ... }
    fn walk_mut(&mut self, hlir: &'a mut HLIR) { ... }

}

Provides an interface from traversing the HLIR tree.

Functions named super_ are default implementations that traverse the tree further, and are not meant to be overriden. They should be called during the corresponding visit_ function implementations to keep traversing the tree further, you can also omit this under conditions if you do not want traversal to continue.

The super_ functions will call the appropriate visit_ functions for child nodes, which should then call the super_ functions to continue traversal, if desired.

By default all visit_ functions call their corresponding super_ functions, so only the kinds of nodes you want to handle need to be overridden/implemented.

Implementations should be provided by overriding the visit_ functions for the specified types you want.

To start traversal, the function walk_mut should be called on this trait with a reference to the HLIR.

Safety

This is safe as long as implementations do NOT store the references obtained in the functions and are only used within the scope of the trait function bodies.

Provided Methods

fn super_root_mut(&mut self, hlir: &mut HLIRDisjointMut<'a>)

Default implementation for visiting the root of the HLIR. This visits the root module if it exists.

fn super_module_mut( &mut self, module: &mut Module, hlir: &mut HLIRDisjointMut<'a>, )

Default implementation for visiting a module. This visits all items within the module.

fn super_item_mut(&mut self, item: &mut Item, hlir: &mut HLIRDisjointMut<'a>)

Default implementation for visiting an item. This dispatches to the appropriate visit function based on the item kind.

fn super_impl_item_mut( &mut self, item: &mut Item, hlir: &mut HLIRDisjointMut<'a>, )

Default implementation for visiting an item within an impl block. This simply calls the super_item_mut function.

fn super_use_mut( &mut self, _use_info: &mut UsePath, _hlir: &mut HLIRDisjointMut<'a>, )

Default implementation for visiting a use path. This does nothing by default.

fn super_impl_mut( &mut self, impl_info: &mut Impl, hlir: &mut HLIRDisjointMut<'a>, )

Default implementation for visiting an impl block. This visits all items within the impl block.

fn super_function_param_mut( &mut self, _parameter: &mut Parameter, _hlir: &mut HLIRDisjointMut<'a>, )

Default implementation for visiting a function parameter. This does nothing by default.

fn super_function_mut( &mut self, function: &mut Function, hlir: &mut HLIRDisjointMut<'a>, )

Default implementation for visiting a function. This visits the function parameters and the function body block.

fn super_enum_mut( &mut self, _enumeration: &mut Enum, _hlir: &mut HLIRDisjointMut<'a>, )

Default implementation for visiting an enum. This does nothing by default.

fn super_struct_mut( &mut self, _structure: &Struct, _hlir: &mut HLIRDisjointMut<'a>, )

Default implementation for visiting a struct. This does nothing by default.

fn super_block_mut(&mut self, block: &Block, hlir: &mut HLIRDisjointMut<'a>)

Default implementation for visiting a block. This visits all statements within the block.

fn super_statement_mut( &mut self, statement: &mut Statement, _parent_block: &Block, hlir: &mut HLIRDisjointMut<'a>, )

Default implementation for visiting a statement. This dispatches based on the statement kind, either visiting a let statement, and item or an expression.

fn super_let_statement_mut( &mut self, _id: HirId, let_statement: &mut LetStatement, hlir: &mut HLIRDisjointMut<'a>, )

Default implementation for visiting a let statement. This visits the initial value expression if it exists.

fn super_constant_mut( &mut self, constant: &mut Constant, hlir: &mut HLIRDisjointMut<'a>, )

Default implementation for visiting a constant. This visits the expression of the constant.

fn super_expr_mut(&mut self, expr: &mut Expr, hlir: &mut HLIRDisjointMut<'a>)

Default implementation for visiting an expression. This dispatches based on the expression kind, visiting child expressions/blocks as necessary.

fn super_path_mut( &mut self, _id: HirId, _path: &mut RefPath, _hlir: &mut HLIRDisjointMut<'a>, )

Default implementation for visiting a path. This does nothing by default.

fn visit_expr_by_id_mut( &mut self, expr_id: HirId, hlir: &mut HLIRDisjointMut<'a>, )

Helper function to call visit_expr_mut by using a given HirId.

fn visit_block_by_id_mut( &mut self, block_id: HirId, hlir: &mut HLIRDisjointMut<'a>, )

Helper function to call visit_block_mut by using a given HirId.

fn visit_path_mut( &mut self, hir_id: HirId, path: &mut RefPath, hlir: &mut HLIRDisjointMut<'a>, )

Called for every path in the HLIR.

fn visit_expr_mut(&mut self, expr: &mut Expr, hlir: &mut HLIRDisjointMut<'a>)

Called for every expression in the HLIR.

fn visit_constant_mut( &mut self, constant: &mut Constant, hlir: &mut HLIRDisjointMut<'a>, )

Called for every constant in the HLIR.

fn visit_let_statement_mut( &mut self, id: HirId, let_statement: &mut LetStatement, hlir: &mut HLIRDisjointMut<'a>, )

Called for every let statement in the HLIR.

fn visit_statement_mut( &mut self, statement: &mut Statement, parent_block: &Block, hlir: &mut HLIRDisjointMut<'a>, )

Called for every statement in the HLIR.

fn visit_block_mut(&mut self, block: &mut Block, hlir: &mut HLIRDisjointMut<'a>)

Called for every block in the HLIR.

fn visit_struct_mut( &mut self, structure: &mut Struct, hlir: &mut HLIRDisjointMut<'a>, )

Called for every struct in the HLIR.

fn visit_enum_mut( &mut self, enumeration: &mut Enum, hlir: &mut HLIRDisjointMut<'a>, )

Called for every enum in the HLIR.

fn visit_function_mut( &mut self, function: &mut Function, hlir: &mut HLIRDisjointMut<'a>, )

Called for every function in the HLIR.

fn visit_function_param_mut( &mut self, parameter: &mut Parameter, hlir: &mut HLIRDisjointMut<'a>, )

Called for every function parameter in the HLIR.

fn visit_impl_mut( &mut self, impl_info: &mut Impl, hlir: &mut HLIRDisjointMut<'a>, )

Called for every impl block in the HLIR.

fn visit_use_mut( &mut self, use_info: &mut UsePath, hlir: &mut HLIRDisjointMut<'a>, )

Called for every use path in the HLIR.

fn visit_item_mut(&mut self, item: &mut Item, hlir: &mut HLIRDisjointMut<'a>)

Called for every item in the HLIR.

fn visit_impl_item_mut( &mut self, item: &mut Item, hlir: &mut HLIRDisjointMut<'a>, )

Called for every item within an impl block in the HLIR.

fn visit_module_mut( &mut self, module: &mut Module, hlir: &mut HLIRDisjointMut<'a>, )

Called for every module in the HLIR.

fn visit_root_mut(&mut self, hlir: &mut HLIRDisjointMut<'a>)

Entry point for visiting the HLIR.

fn walk_mut(&mut self, hlir: &'a mut HLIR)

Entry point for walking the HLIR mutably.