kitlang::intermediate

Module mir

The MIR module provides a control flow based IR suitable for interpretation, further optimization, and eventually lowering further into machine code or bytecode.

The MIR represents function bodies as collections of BasicBlocks containing statements and instructions on what to do when exiting the block (ExitDirective). Each function Body maintains local variables and the collection of BasicBlocks that make up the function’s control flow.

This representation simplifies operations to assignments, branches, and calls, making it much closer to how a computer would execute the code, and is quite far removed from the original or even HIR forms.

At this stage, things such as identifiers, etc, have been fully removed and everything operates on essentially indexes of local variables, or an index (will eventually be a byte offset) into a local variable for field accesses.

Modules

lowerer 🔒

Structs

BasicBlock

BasicBlockId

Body

ExitDirective

LocalDefinition

LocalId

MIR

Statement

Enums

ADTKind

AssignTarget

Defines a slot that is the target for a value to be assigned or copied into. This can either be a LocalId or a Field defined in a LocalId.

BlockExitKind

CastKind

LocalInfo

Operand

RValue

StatementKind

Functions

lower_hir_to_mir

Errors