kitlang::intermediate::resolver::associated_references

Struct AssociatedReferenceMapper

struct AssociatedReferenceMapper<'a> {
    pub meta: &'a mut ProgramMetaData,
    pub global_functions: HashMap<String, OwnerDefId>,
    pub path_stack: Vec<IdentPath>,
    pub impl_path_lut: Vec<(OwnerDefId, IdentPath)>,
    pub stage1_complete: bool,
    pub errors: Vec<ResolverError>,
    pub resolution_failures: Vec<(HirId, SpannedIdentPath, ResolutionFailure)>,
}

Fields
meta: &'a mut ProgramMetaDataglobal_functions: HashMap<String, OwnerDefId>

Tracks all defined global functions, for quick lookup during resolution. We check this map first when resolving single-segment paths.

path_stack: Vec<IdentPath>

Stack of identifier paths representing the current scope during traversal. This is used if we need to temporarily change scope while visiting nested items, before reverting back to the previous scope.

impl_path_lut: Vec<(OwnerDefId, IdentPath)>

Lookup table of all impl items encountered during the first pass, so we can defer visiting them until later.

stage1_complete: bool

Indicates whether the first stage of mapping (building namespaces) is complete.

errors: Vec<ResolverError>

Collected resolver errors encountered during mapping.

resolution_failures: Vec<(HirId, SpannedIdentPath, ResolutionFailure)>

Collected resolution failures for unresolved references, to be reported after resolution. Each entry contains the HirId of the item that failed, the SpannedIdentPath it is trying to access, and the ResolutionFailure reason.

Implementations

impl<'a> AssociatedReferenceMapper<'a>

pub fn new(meta: &'a mut ProgramMetaData) -> Self

pub fn reset_path_to(&mut self, ident_path: IdentPath)

Resets the current path stack to only contain ident_path.

pub fn reset_path(&mut self)

Resets the current path stack to be empty, i.e. the root namespace.

fn with_pushed_segment<F>(&mut self, segment: &str, f: F)
where F: FnOnce(&mut Self),

Executes f with segment temporarily pushed onto the current path. Ensures the segment is popped after f completes.

fn with_pushed_path<F>(&mut self, path: IdentPath, f: F)
where F: FnOnce(&mut Self),

Executes f with path temporarily pushed onto the path stack. Ensures the path is popped after f completes.

fn inject_builtin_definitions(&mut self)

Injects builtin type definitions into the root Namespace.

fn resolve_uses_in_namespace( root_namespace: &Namespace, namespace: &mut Namespace, current_path: &IdentPath, errors: &mut Vec<ResolverError>, )

Recursively resolves all use items in the given Namespace by looking up their target definitions in the root_namespace and cloning them into the current Namespace.

pub fn map_references(&mut self, hlir: &mut HLIR) -> ResolveResult<()>

Entrypoint function to map all associated references in the given HLIR. This will build the Namespace tree and map, then resolve all associated references,

fn find_last_enclosing_path(&self, path: &IdentPath) -> IdentPath

Finds the last enclosing scope path for the given path. Where an enclosing scope is something like a module or function that defines a ‘border’ of accessibility.

fn search_backtracking_for_definition( &self, ident_path: &IdentPath, base_path: IdentPath, ) -> Option<(IdentPath, ResolvedID)>

Searches for the definition of ident_path starting from base_path,

fn validate_local_access( &self, target_path: &IdentPath, base_path: &IdentPath, ) -> bool

Validates that if the target path is marked as local (private), it is being accessed from within its enclosing scope.

Returns

true if access is valid.

false if access violates local visibility rules.

pub fn search_for_definition( &self, ident_path: &IdentPath, base_path: &IdentPath, ) -> Result<ResolvedID, ResolutionFailure>

Searches for the definition of ident_path starting from base_path. This function performs backtracking to find the definition in enclosing scopes, while also enforcing visibility and local access rules.

Returns

Ok(ResolvedID) if the definition is found and accessible.

Err(ResolutionFailure) if the definition is not found or inaccessible.

impl AssociatedReferenceMapper<'_>

pub fn current_path(&self) -> &IdentPath

Get a reference to the current identifier path from the top of the path stack.

pub fn current_path_mut(&mut self) -> &mut IdentPath

Get a mutable reference to the current identifier path from the top of the path stack.

pub fn push_to_current_path(&mut self, s: &str)

Push a new segment onto the end of the current identifier path.

Example

If our current path is ::example and we push inner_module, our new path will be ::example::inner_module.

pub fn pop_from_current_path(&mut self)

Pop the last segment from the end of the current identifier path.

Example

If our current path is ::example::inner_module and we pop, our new path will be ::example.

pub fn push_stack(&mut self, path: IdentPath)

Push a new identifier path onto the top of the path stack.

pub fn pop_stack(&mut self)

Pop the top identifier path from the path stack.

impl AssociatedReferenceMapper<'_>

fn should_be_local(&self, path: &IdentPath) -> bool

Determines if the given path should be marked as local based on its namespace. A path is considered local if it is defined within a local scope or is a function

fn get_namespace(&self, path: &IdentPath) -> Option<&Namespace>

Get a reference to the Namespace defined at a given path, if it exists.

fn get_namespace_mut(&mut self, path: &IdentPath) -> Option<&mut Namespace>

Get a mutable reference to the Namespace defined at a given path, if it exists.