kitlang::intermediate::resolver

Module errors

The purpose of this module is to define error types for the resolution phase of the compiler. Provides structured error representations for various resolution failures, including unresolved references, inaccessible items, and general diagnostic messages.

Macros

push_resolve_err 🔒

resolve_err 🔒

Structs

ResolverError

Represents an error while resolving references/symbols after in the intermediate representation.

UnresolvedReference

Represents a single unresolved reference in the intermediate representation. It contains the path of the reference, its associated HirId, and the reason for the failure.

UnresolvedReferences

Represents a collection of unresolved references in the intermediate representation. Used for error handling, to bundle up multiple resolution failures into one error value.

Enums

ResolutionFailure

Represents the reason for a resolution failure.

ResolverErrorKind

Type Aliases

ResolveResult