kitlang::intermediate::resolver

Module locals

The purpose of this module is to manage local scopes during the resolution phase of the compiler. A local scope is a scope in which local variables are defined and can be resolved. This is usually within functions, blocks, or other constructs that introduce a new scope.

This file is NOT responsible for managing scopes between modules, other functions, etc. That is handled by the associated_references resolver.

This module is exclusively focused on local variable resolution, and ensuring that variables are correctly scoped and resolved within their respective local contexts. All other references are handled elsewhere.

Structs

LocalScope 🔒

Represents a local scope in the resolver, maintaining variable definitions and their resolved IDs.

ScopeResolver 🔒

A resolver that manages local scopes and resolves local variable references in the HLIR. We keep track of resolver state in this struct, and use it to implement the HLIR visitor traits. In this case, the scope resolver only requires one mutable pass over the HLIR to resolve local references.

Functions

resolve_scope_paths

Resolves local variable references in the given HLIR. This function performs a mutable walk over the HLIR, resolving local variable references.