kitlang::intermediate::resolver::locals

Struct LocalScope

struct LocalScope {
    pub scope_ident: Option<String>,
    pub definitions: HashMap<String, ResolvedID>,
    pub parent: Option<Box<Self>>,
}

Represents a local scope in the resolver, maintaining variable definitions and their resolved IDs.

These local scopes may be nested inside a parent scopes, forming a hierarchy of scopes. Expressions inside scopes may only access variables defined in their own scope or in parent scopes, not variables in child scopes.

Fields
scope_ident: Option<String>

An optional identifier for the scope, useful for debugging or error messages.

definitions: HashMap<String, ResolvedID>

A mapping from variable names to their resolved IDs within this scope.

parent: Option<Box<Self>>

An optional parent scope, allowing for nested scopes.

Implementations

impl LocalScope

pub fn new(scope_ident: Option<String>) -> Self

Create a new local scope with an optional identifier.

pub fn new_with_parent( scope_ident: Option<String>, parent: Option<Box<Self>>, ) -> Self

Create a new local scope with an optional identifier and a parent scope.

pub fn child_scope(&self, scope_ident: Option<String>) -> Self

Create a child scope of the current scope with an optional identifier.

pub const fn is_root_scope(&self) -> bool

Check if this scope is the root scope (i.e., has no parent).

pub fn add_definition_unique(&mut self, name: &str, id: ResolvedID) -> bool

Add a new definition, ensuring that the name is unique within this scope.

Returns

true if the definition was added successfully, false if the name already exists.

pub fn add_definition_overwrite(&mut self, name: &str, id: ResolvedID) -> bool

Add a new definition, redefining the value that was already there, if exists.

Returns

true if the definition overwrote an existing one, false if it was newly added.

pub fn add_definition_result( &mut self, name: &str, id: ResolvedID, ) -> ResolveResult<()>

Add a new definition, returning an error if the name already exists in this scope.

Returns

Ok(()) if the definition was added successfully

Err(ResolverError) if the name already exists.

pub fn find_definition(&self, name: &str) -> Option<ResolvedID>

Find a definition in the current scope or any parent scopes.

Returns

Some(ResolvedID) if the definition is found.

None if the definition is not found.