kitlang::intermediate::resolver::locals

Struct ScopeResolver

struct ScopeResolver {
    pub scope: Vec<LocalScope>,
    pub errors: Vec<ResolverError>,
}

A resolver that manages local scopes and resolves local variable references in the HLIR. We keep track of resolver state in this struct, and use it to implement the HLIR visitor traits. In this case, the scope resolver only requires one mutable pass over the HLIR to resolve local references.

Fields
scope: Vec<LocalScope>errors: Vec<ResolverError>

Collected resolver errors during the resolution process.

Implementations

impl ScopeResolver

pub fn push_scope(&mut self, local_scope: LocalScope)

Push a new local scope onto the scope stack.

pub fn pop_scope(&mut self)

Pop the current local scope from the scope stack.

pub fn current_scope(&self) -> &LocalScope

Get a reference to the current local scope.

pub fn current_scope_mut(&mut self) -> &mut LocalScope

Get a mutable reference to the current local scope.

pub fn push_child_scope(&mut self, scope_ident: Option<String>)

Create a new child scope of the current scope, with an optional identifier, and push it onto the scope stack.

fn with_scope<F>(&mut self, scope_ident: Option<String>, f: F)
where F: FnOnce(&mut Self),

Executes f within a new local scope with an optional identifier. The new scope is pushed onto the scope stack before executing f, and popped off the stack after f completes.

fn with_child_scope<F>(&mut self, scope_ident: Option<String>, f: F)
where F: FnOnce(&mut Self),

Executes f within a new child scope of the current scope with an optional identifier. The new scope is pushed onto the scope stack before executing f, and popped off the stack after f completes.

pub fn resolve(&mut self, hlir: &mut HLIR) -> ResolveResult<()>

Entrypoint to resolve local variable references in the given HLIR. This function performs a mutable walk over the HLIR, resolving local variable references.

Returns

Ok(()) if all local variable references were resolved successfully.

Err(ResolverError) if there were any resolution errors.

Errors

This function will return an error if any local variable references could not be resolved. The returned error may contain multiple resolution errors if there were multiple failures.