kitlang::intermediate::resolver::types

Function resolve_types

pub fn resolve_types(
    hlir: &mut HLIR,
    meta: &mut ProgramMetaData,
) -> ResolveResult<()>

Resolves all types referenced by a path to a TypeID in the given HLIR. This function performs a mutable walk over the HLIR, resolving type references.

Returns

Ok(()) if all type references were resolved successfully.

Err(ResolverError) if there were any resolution errors.

Errors

This function will return an error if any type references could not be resolved. The returned error may contain multiple resolution errors if there were multiple failures.