kitlang::intermediate::resolver::verifier

Struct UnresolvedReferenceChecker

struct UnresolvedReferenceChecker {
    pub unresolved_references: Vec<UnresolvedReference>,
}

Visitor that checks for unresolved references in the HLIR. It collects all unresolved references found during the visit.

Fields
unresolved_references: Vec<UnresolvedReference>

All found unresolved references.

Implementations

impl UnresolvedReferenceChecker

pub fn verify_references(self, hlir: &HLIR) -> ResolveResult<()>

Verifies all references in the given HLIR. Visits all path nodes and collects any that are still unresolved.

Returns

Ok(()) if all references are resolved.

Err(ResolverError) if there are any unresolved references.

Errors

This function will return an error if there are any unresolved references in the HLIR remaining. The returned error will contain all unresolved references found.