kitlang::intermediate::type_check

Function run_type_checker

pub fn run_type_checker(
    hlir: &mut HLIR,
    meta: &mut ProgramMetaData,
) -> LowerResult<()>

This function will run the type checker stage on the resolved HIR output. This will validate that all type rules are followed, as well as fill in any types that must be inferred from context.

Errors

This function will return an error if any part of the HIR fails to pass type checking. The returned error will contain a diagnostic message indicating the nature and location of any failures or incompatibilities as well as where in the source code they occurred. This function can return multiple diagnostics messages/errors at once, to provide a more comprehensive overview of all type checking issues in the HIR.

Returns

Ok(()) if the type checking passes with no errors.

Err(LoweringError) if there are type checking failures, containing details of the failures.