kitlang::intermediate::type_check

Struct TypeChecker

struct TypeChecker<'a> {
    pub meta: &'a mut ProgramMetaData,
    pub should_infer: bool,
    pub errors: Vec<TypeCheckFail>,
    pub return_type_stack: Vec<Type>,
}

Type checker that validates types in the HLIR after resolution.

Fields
meta: &'a mut ProgramMetaDatashould_infer: bool

Whether to perform type inference where types are not explicitly provided.

errors: Vec<TypeCheckFail>

Collected type checking failures.

return_type_stack: Vec<Type>

Stack to track the expected return types in nested function contexts.

Implementations

impl<'a> TypeChecker<'a>

pub const fn new(meta_data: &'a mut ProgramMetaData, should_infer: bool) -> Self

impl TypeChecker<'_>

pub fn push_current_return_type(&mut self, t: Type)

Pushes the current expected return type onto the stack.

pub fn pop_current_return_type(&mut self)

Pops the current expected return type from the stack.

pub fn current_expected_return(&self) -> Option<Type>

Retrieves the current expected return type from the top of the stack, if it is set.

pub fn try_type_name(&self, ty: impl Into<Type>) -> Option<String>

Attempts to get a human-readable type name for the given Type.

pub fn type_name(&self, ty: impl Into<Type>) -> String

Gets a human-readable type name for the given Type, defaulting to "??" if it cannot be determined.

impl TypeChecker<'_>

fn resolved_type( &self, id: HirId, t: &Type, hlir: &HLIRDisjointMut<'_>, ) -> TypeResult<KitTy>

Ensures that the given type is resolved.

Returns

Ok(KitTy) if the type is resolved.

Err(TypeCheckFail) if the type is unresolved.

impl TypeChecker<'_>

fn get_func_sig_by_call_expr_id( id: HirId, hlir: &mut HLIRDisjointMut<'_>, ) -> TypeResult<(Type, Vec<Type>)>

Retrieves the function signature (return type and parameter types) for a function call expression by its HirId.

Returns

Ok((Type, Vec<Type>)) containing the return type and parameter types if successful.

Err(TypeCheckFail) if there was an error retrieving the function signature.

fn get_type_of_function_param( func_id: OwnerDefId, expr_id: HirId, param_index: u32, hlir: &mut HLIRDisjointMut<'_>, ) -> TypeResult<Type>

Retrieves the type of a function parameter by its function OwnerDefId and parameter index. Also takes the expression HirId for error reporting.

Returns

Ok(Type) if the parameter type is found.

Err(TypeCheckFail) if there was an error retrieving the parameter type.

impl TypeChecker<'_>

fn validate_return_value( &self, id: HirId, t: Type, hlir: &HLIRDisjointMut<'_>, ) -> TypeResult<Type>

Validates that the given return type matches the expected return type for the current function context. This is called when a root block is being type checked on it’s final expression.

Returns

Ok(Type) if the return type matches the expected type.

Err(TypeCheckFail) if there was a type mismatch.

fn validate_return_value_by_id( &mut self, expr_id: HirId, hlir: &mut HLIRDisjointMut<'_>, ) -> TypeResult<Type>

Validates that the return type of the expression with the given HirId matches the expected return type.

Returns

Ok(Type) if the return type matches the expected type.

Err(TypeCheckFail) if there was a type mismatch, or other error while evaluating the expression type.

fn eval_non_expr_hir_id( id: HirId, hlir: &mut HLIRDisjointMut<'_>, ) -> TypeResult<Type>

Evaluates the type of a non-expression HLIR node (e.g., parameter or statement) by its HirId.

Returns

Ok(Type) if the non-expression type is successfully evaluated.

Err(TypeCheckFail) if there was an error during type evaluation.

fn eval_expr_type_by_id( &mut self, id: HirId, hlir: &mut HLIRDisjointMut<'_>, ) -> TypeResult<Type>

Evaluates the type of an expression by its HirId, updating the type map accordingly.

Returns

Ok(Type) if the expression type is successfully evaluated.

Err(TypeCheckFail) if there was an error during type evaluation.

fn eval_expr_type_by_id_expected( &mut self, id: HirId, hlir: &mut HLIRDisjointMut<'_>, expected: &Type, ) -> TypeResult<Type>

Evaluates the type of an expression by its HirId, updating the type map accordingly. Also returns whether the expression is a literal.

Returns

Ok((Type, is_literal)) if the expression type is successfully evaluated.

Err(TypeCheckFail) if there was an error during type evaluation.

fn eval_expr_type( &mut self, expr: &Expr, hlir: &mut HLIRDisjointMut<'_>, expected: Option<&Type>, ) -> TypeResult<Type>

Evaluates the type of an expression node.

Returns

Ok(Type) if the expression type is successfully evaluated.

Err(TypeCheckFail) if there was an error during type evaluation.

fn eval_and_infer_let_statement( &mut self, id: HirId, let_statement: &mut LetStatement, hlir: &mut HLIRDisjointMut<'_>, ) -> TypeResult<Type>

Evaluates and infers the type of a let statement, updating the statement’s type if necessary.

Returns

Ok(Type) if the let statement type is successfully evaluated or inferred.

Err(TypeCheckFail) if there was an error during type evaluation or inference.

fn eval_statement_type( &mut self, statement: &mut Statement, _parent_block: &Block, hlir: &mut HLIRDisjointMut<'_>, ) -> TypeResult<Type>

Evaluates the type of a statement within a block.

Returns

Ok(Type) if the statement type is successfully evaluated.

Err(TypeCheckFail) if there was an error during type evaluation.

fn eval_block_type_by_id( &mut self, id: HirId, hlir: &mut HLIRDisjointMut<'_>, ) -> TypeResult<Type>

Evaluates the type of a block by its HirId.

Returns

Ok(Type) if the block type is successfully evaluated.

Err(TypeCheckFail) if there was an error during type evaluation.

fn is_id_return_statement( id: HirId, hlir: &mut HLIRDisjointMut<'_>, ) -> TypeResult<bool>

Checks if the statement with the given HirId is a return statement.

Returns

Ok(bool) indicating whether the statement is a return statement.

Err(TypeCheckFail) if there was an error retrieving the statement.

fn eval_block_type( &mut self, block: &Block, hlir: &mut HLIRDisjointMut<'_>, ) -> TypeResult<Type>

Evaluates the type of a block.

Returns

Ok(Type) if the block type is successfully evaluated.

Err(TypeCheckFail) if there was an error during type evaluation.

fn eval_owner_def( &mut self, owner_def_id: OwnerDefId, hlir: &mut HLIRDisjointMut<'_>, ) -> TypeResult<()>

Evaluates an owner definition (e.g., function, use statement). This is a dispatch for type checking the internals of various owner definitions.

Returns

Ok(()) if the owner definition is successfully evaluated.

Err(TypeCheckFail) if there was an error during evaluation.