kitlang::interpreter::mir_interpreter

Enum Value

pub enum Value {
    Unit,
    Integer(i64),
    UnsignedInteger(u64),
    Float(f64),
    String(String),
    Boolean(bool),
    Ref(AssignTarget),
    ADT(ADTValueKind),
    Tuple(Vec<Value>),
    Array(Vec<Value>),
}

Variants

Unit

Integer(i64)

UnsignedInteger(u64)

Float(f64)

String(String)

Boolean(bool)

Ref(AssignTarget)

ADT(ADTValueKind)

Tuple(Vec<Value>)

Array(Vec<Value>)

Implementations

impl Value

pub fn from_literal(lit: Literal) -> Self

pub fn repr_string(&self) -> String

pub const fn int(&self) -> Option<i64>

pub const fn int_coerce(&self) -> Option<i64>

pub const fn uint(&self) -> Option<u64>

pub const fn uint_coerce(&self) -> Option<u64>

pub const fn float(&self) -> Option<f64>

pub fn string(&self) -> Option<String>

pub fn str_ref(&self) -> Option<&str>

pub const fn bool(&self) -> Option<bool>

pub const fn are_matching_types(&self, other: &Self) -> bool

pub fn perform_unary_op(&self, op: UnaryOpKind) -> Option<Self>

pub fn perform_binary_op(&self, rhs: &Self, op: BinaryOpKind) -> Option<Self>

Perform a binary operation between this value and another value.

Panics

Panics if the values are either Ref or ADT variants as they are not yet implemented.

pub const fn perform_increment(&self) -> Option<Self>

pub const fn as_index_usize(&self) -> Option<usize>

Converts the value to a usize index if possible.

pub fn index(&self, index: usize) -> Option<&Self>

pub fn index_mut(&mut self, index: usize) -> Option<&mut Self>

impl Value

pub const fn is_unit(&self) -> bool

pub const fn is_reference(&self) -> bool

pub const fn is_adt(&self) -> bool

pub const fn is_tuple(&self) -> bool

pub const fn is_array(&self) -> bool