kitlang::interpreter::mir_interpreter

Struct InterpreterState

pub struct InterpreterState {
    pub entry: OwnerDefId,
    pub native_functions: HashMap<String, Box<KitlangMIRNativeFn>>,
    pub ctx: ExecutionContext,
}

Fields
entry: OwnerDefIdnative_functions: HashMap<String, Box<KitlangMIRNativeFn>>ctx: ExecutionContext
Implementations

impl InterpreterState

const DEFAULT_ENTRY_NAME: &str = "main"

fn find_entry_point(program: &ProgramType) -> Option<OwnerDefId>

pub fn new(program: &ProgramType) -> Option<Self>

impl InterpreterState

pub fn register_native_function<F: IntoMIRKitlangFn>( &mut self, name: &str, func: F, )

pub fn call_native_function( &mut self, name: &str, args: &[Value], ) -> Result<Value, InterpreterError>

pub fn execute_from_entry( &mut self, program: &ProgramType, ) -> Result<Value, InterpreterError>

impl InterpreterState

pub fn execute_function( &mut self, program: &ProgramType, id: OwnerDefId, args: &[Value], ) -> Result<Value, InterpreterError>

Execute a function by its OwnerDefId.

Errors

Returns an error if the function is not found or if there is an error during execution. The returned error contains diagnostic information about the error.

fn execute_kit_function( &mut self, program: &ProgramType, body: &Body, args: &[Value], ) -> Result<Value, InterpreterError>

fn eval_operand(&self, operand: &Operand) -> Result<Value, InterpreterError>

fn eval_rvalue(&self, rvalue: &RValue) -> Result<Value, InterpreterError>

pub fn perform_assignment( &mut self, target: AssignTarget, new_value: Value, ) -> Result<(), InterpreterError>

pub fn perform_cast( &self, value: &Value, target_type: CastKind, ) -> Option<Value>

pub fn deref_value(&self, value: Value) -> Result<Value, InterpreterError>

pub fn execute_statement( &mut self, statement: &Statement, ) -> Result<(), InterpreterError>