kitlang

Module lexer

The lexer module transforms source code strings into streams of Tokens.

Provides a way to peek and consume characters from the source code string as a Stream like structure, as well as skipping whitespace, etc.

The main API for this module are the tokenise and tokenise_stripped functions, which return iterators that will lazily parse tokens from the source string, with tokenise_stripped ignoring comments and documentation tokens.

Structs

CodeCursor

Iterator over a character sequence, supports peeking and reading from the sequence.

Lexer 🔒

Holds the state of the lexing process, contains implementation of the lexer.

Constants

EOF_CHAR 🔒

End of file.

Functions

is_comment_sequence 🔒

Returns true if the character sequence describes a comment or doc block.

is_comment_token_kind 🔒

Returns true if the token kind is a comment, or documentation token.

is_newline_char 🔒

Returns true if the char is considered a newline char ('\r' or '\n')

is_whitespace

Returns true if c is considered a ‘whitespace’ character.

tokenise

Returns an iterator that lazily parses tokens from the source string.

tokenise_stripped

Returns an iterator that lazily parses tokens from the source string, ignoring ‘unneccesary’ tokens such as comments and documentation.

Type Aliases

CodeCursorIndex

Type used to describe an index of a CodeCursor.