kitlang::lexer

Struct CodeCursor

pub struct CodeCursor<'a> {
    length: CodeCursorIndex,
    chars: Chars<'a>,
    previous: char,
}

Iterator over a character sequence, supports peeking and reading from the sequence.

Fields
length: CodeCursorIndex

Length of the input string in characters.

chars: Chars<'a>

Iterator over code-points of a source string.

previous: char

Previous character to be read.

Implementations

impl<'a> CodeCursor<'a>

pub fn new(input: &'a str) -> Self

Construct a new code character iterator with a provided input string.

pub fn as_str(&self) -> &'a str

Returns the slice from the current cursor position to the end of the original underlying data of the CodeCursor.

pub fn position(&self) -> CodeCursorIndex

The current position of the cursor in the string.

pub fn remaining(&self) -> CodeCursorIndex

The amount of characters remaining.

pub const fn prev(&self) -> char

Get the last ‘consumed’ symbol.

Returns

The previous symbol to be consumed by the stream, or EOF if there is no valid previous.

pub fn peek(&self) -> char

Peek the next or ‘current’ character, without consuming it. (Without advancing the cursor).

Returns

The character at the current cursor position, or the EOF char if the position is invalid.

pub fn peek_opt(&self) -> Option<char>

Peek the next or ‘current’ character, without consuming it. (Without advancing the cursor).

pub fn peek_second(&self) -> char

Peek at the character after the current character, without consuming (without advancing the cursor).

pub fn peek_third(&self) -> char

Peek at the character two characters after the current character, without consuming (without advancing the cursor).

pub fn peek_at(&self, offset: CodeCursorIndex) -> Option<char>

Peek forward a specified number of characters, this does not advance the cursor.

Returns

The character at the specified offset or None if the iterator could not be advanced to that point.

pub fn compare_next(&self, expected: &str) -> bool

Compare the string representation of the cursor’s remaining data to check if the next len(expected) characters are equal to expected.

pub fn consume(&mut self) -> Option<char>

Consume the current cursor position, advancing the cursor by one.

Returns the next character in the sequence if valid, None otherwise.

pub fn consume_while( &mut self, predicate: impl FnMut(char) -> bool, ) -> CodeCursorIndex

Advance the cursor over characters until predicate returns false or the EOF is reached.

Returns

How many characters were consumed.

Panics

Panics if consume fails during execution, I.e. if eof is reached unexpectedly.

pub fn consume_until_str(&mut self, sequence: &str) -> Option<CodeCursorIndex>

Continue to consume characters up until a specified character sequence is found. This function will not consume the start of the sequence, it peeks, then consumes if the sequence/pattern is not found.

Returns

The number of characters consumed before the sequence was found or eof was encountered.

pub fn consume_until_sequence( &mut self, sequence: &[char], ) -> Option<CodeCursorIndex>

Panics

Panics if consume fails during execution, I.e. if eof is reached unexpectedly.

pub fn consume_whitespace(&mut self) -> CodeCursorIndex

Advance the cursor over all whitespace characters until a non-whitespace character is found.

Returns

How many characters were consumed.

pub fn consume_newline(&mut self) -> Option<CodeCursorIndex>

Consume the newline character sequence, this can either be '\n' or "\r\n". This function will consume either 1 or 2 characters depending on the line ending of the input. The sequence '\r' with no newline after it is considered invalid.

Returns

None if the cursor was not at a newline char, or the number of characters consumed if it was.

pub fn consume_expect( &mut self, expected_chars: &[char], ) -> Option<CodeCursorIndex>

Consume the current cursor position, advancing the cursor by one for each character in expected_chars, ensures the consumed character was equal to the expected character.

Returns

The amount of characters consumed, or None if one of the expected chars was not found, or if eof was reached.

pub fn consume_expect_str(&mut self, expected: &str) -> Option<CodeCursorIndex>

Consume the current cursor position, advancing the cursor by one for each character in expected, checking if the consumed character was equal to the corresponding char in expected.

Returns

The amount of characters consumed, or None if one of the expected chars was not equal, or if eof was reached.

pub fn is_eof(&self) -> bool

Returns true if the cursor as at the end of the file.