kitlang::lexer

Struct Lexer

struct Lexer<'a> {
    source_string: &'a str,
    cursor: CodeCursor<'a>,
}

Holds the state of the lexing process, contains implementation of the lexer.

Fields
source_string: &'a strcursor: CodeCursor<'a>
Implementations

impl<'a> Lexer<'a>

pub fn new(source_string: &'a str) -> Self

Construct a new lexer state from an input string.

pub fn read_token(&mut self) -> Option<Token>

Read a token from the current cursor position.

fn substr(&self, start: CodeCursorIndex, end: CodeCursorIndex) -> &'a str

Returns a substring of the source_string starting at the start index, and ending exclusively at the end index.

fn substr_inclusive( &self, start: CodeCursorIndex, end: CodeCursorIndex, ) -> &'a str

Returns a substring of the source_string starting at the start index, and ending inclusively at the end index.

impl Lexer<'_>

fn documentation(&mut self) -> Option<Token>

Parses a documentation block from the cursor position.

Details

Parses every line that starts with ‘///’ until the next line with a non-documentation or comment token is found, as a documentation token.

fn line_comment(&mut self) -> Option<Token>

Parses a line comment from the cursor.

Details

Parse everything after '//' until a newline as a comment.

fn block_comment(&mut self) -> Option<Token>

Parses a block comment from the cursor.

Details

Parses everything between '/*' and '*/ as a comment.

fn punctuation(&mut self) -> Option<Token>

Parses a punctuation symbol from the cursor.

fn identifier(&mut self) -> Token

Parses an identifier from the cursor.

fn numeric_literal(&mut self) -> Option<Token>

Parses a numeric literal from the cursor.

Details

All support underscores (‘_’) for delimiting, underscores are to be read as non-influential characters. Can be of the form:

Binary (‘0b01010101’)

Hexadecimal (‘0xFA’)

Octal (‘0o80’)

Decimal (‘820’) and (2.1023)

E-notation (‘2.1e6’)

fn string_literal(&mut self) -> Option<Token>

Parses a string literal from the cursor.