kitlang::parser

Function parse_from_source

pub fn parse_from_source(source: &str) -> Result<ASTRoot, ParseError>

Parse AST from a source code string.

Note

This function will collect the iterator into a Vec<Token>.

This function also strips all comments and documentation from the code.

Errors

This function will return an error if any part of parsing the token stream fails. The returned error will contain information about where in the source code the error occurred, and a message about the nature of the error.