kitlang::parser

Struct Parser

pub(crate) struct Parser<'a, 'b> {
    context: &'a ParserContext,
    cursor: TokenCursor<'b>,
}

Contains the state needed to parse an input file, as well as methods to parse different Items, Expressions and Statements.

Fields
context: &'a ParserContextcursor: TokenCursor<'b>
Implementations

impl Parser<'_, '_>

fn is_binary_op(&self, offset: u32) -> Result<Option<BinaryOpKind>, ParseError>

fn is_an_assign_op(&self, offset: u32) -> Result<bool, ParseError>

fn is_double_eq(&self) -> bool

fn is_double_dot(&self) -> bool

fn process_atom( &mut self, atom: Box<Expression>, span_start: u32, ) -> Result<Box<Expression>, ParseError>

pub fn parse_expr_atom(&mut self) -> Result<Box<Expression>, ParseError>

fn parse_expr_atom_with_cast(&mut self) -> Result<Box<Expression>, ParseError>

fn parse_parens_expr(&mut self) -> Result<Box<Expression>, ParseError>

fn parse_array_initialiser(&mut self) -> Result<Box<Expression>, ParseError>

fn parse_range_expr_continued( &mut self, lhs: Box<Expression>, span_start: u32, ) -> Result<Box<Expression>, ParseError>

pub fn parse_expression(&mut self) -> Result<Box<Expression>, ParseError>

pub fn parse_block_as_expression( &mut self, ) -> Result<Box<Expression>, ParseError>

pub fn parse_block_expression(&mut self) -> Result<Box<Block>, ParseError>

fn invalid_literal(&mut self) -> Result<Box<Expression>, ParseError>

fn parse_literal(&mut self) -> Result<Box<Expression>, ParseError>

fn parse_assign_continued( &mut self, lhs: Box<Expression>, span_start: u32, ) -> Result<Box<Expression>, ParseError>

fn parse_unary_op(&mut self) -> Result<Box<Expression>, ParseError>

fn parse_assign_op_continued( &mut self, lhs: Box<Expression>, op_kind: BinaryOpKind, span_start: u32, ) -> Result<Box<Expression>, ParseError>

fn parse_binary_op_continued_ordered( &mut self, lhs: Box<Expression>, min_order: ExpressionOrderBound, span_start: u32, ) -> Result<Box<Expression>, ParseError>

fn parse_binary_op_continued( &mut self, lhs: Box<Expression>, span_start: u32, ) -> Result<Box<Expression>, ParseError>

fn parse_cast_op_continued( &mut self, lhs: Box<Expression>, span_start: u32, ) -> Result<Box<Expression>, ParseError>

fn parse_call_continued( &mut self, lhs: Box<Expression>, span_start: u32, ) -> Result<Box<Expression>, ParseError>

fn parse_if(&mut self) -> Result<Box<Expression>, ParseError>

fn parse_loop(&mut self) -> Result<Box<Expression>, ParseError>

fn parse_for(&mut self) -> Result<Box<Expression>, ParseError>

fn parse_while(&mut self) -> Result<Box<Expression>, ParseError>

fn parse_break(&mut self) -> Result<Box<Expression>, ParseError>

fn parse_continue(&mut self) -> Result<Box<Expression>, ParseError>

fn parse_return(&mut self) -> Result<Box<Expression>, ParseError>

fn parse_array_index( &mut self, lhs: Box<Expression>, span_start: u32, ) -> Result<Box<Expression>, ParseError>

fn parse_member_access( &mut self, lhs: Box<Expression>, span_start: u32, ) -> Result<Box<Expression>, ParseError>

fn parse_member_call( &mut self, lhs: Box<Expression>, span_start: u32, ) -> Result<Box<Expression>, ParseError>

fn parse_self(&mut self) -> Result<Box<Expression>, ParseError>

fn parse_field_initialisation( &mut self, ) -> Result<FieldInitialisation, ParseError>

fn parse_struct_initialiser( &mut self, lhs: SpannedIdentPath, ) -> Result<Box<Expression>, ParseError>

impl Parser<'_, '_>

pub fn parse_item(&mut self) -> Result<Item, ParseError>

pub fn parse_function( &mut self, from_impl_block: bool, ) -> Result<Item, ParseError>

fn parse_struct_field_definition( &mut self, allow_vis_specifier: bool, ) -> Result<StructField, ParseError>

pub fn parse_struct(&mut self) -> Result<Item, ParseError>

fn parse_enum_variant(&mut self) -> Result<EnumVariant, ParseError>

pub fn parse_enum(&mut self) -> Result<Item, ParseError>

pub fn parse_const(&mut self) -> Result<Item, ParseError>

fn parse_impl_item(&mut self) -> Result<Item, ParseError>

pub fn parse_implementation(&mut self) -> Result<Item, ParseError>

fn parse_mod_item(&mut self) -> Result<Item, ParseError>

pub fn parse_use_segment(&mut self) -> Result<Vec<IdentPath>, ParseError>

pub fn parse_use(&mut self) -> Result<Item, ParseError>

pub fn parse_mod(&mut self) -> Result<Item, ParseError>

impl Parser<'_, '_>

pub fn parse_statement(&mut self) -> Result<Statement, ParseError>

pub fn parse_variable_pattern(&mut self) -> Result<VariablePattern, ParseError>

pub fn parse_binding(&mut self) -> Result<BindingPattern, ParseError>

pub fn parse_let_local(&mut self) -> Result<Box<Local>, ParseError>

impl<'a, 'b> Parser<'a, 'b>

pub const fn from_cursor( token_cursor: TokenCursor<'b>, context: &'a ParserContext, ) -> Self

impl Parser<'_, '_>

fn is_double_colon(&self) -> bool

Returns true if the next 2 tokens from the cursor position are Punctuation::Colon.

fn parse_double_colon(&mut self, path_start: u32) -> Result<bool, ParseError>

Parses a double colon "::" or the “path” separator from the cursor position.

This function advances the cursor position by the length of the path separator (2) if the double colon is found.

Returns

Ok(true) if a double colon was found and the cursor was advanced.

Ok(false) if a double colon wasn’t found, but the token at the current cursor position was not a single colon ':', and the cursor was not advanced.

Err(ParseErrorKind::InvalidPath) if the double colon was not found, but a single colon was, indicating a malformed path.

pub fn parse_path_segment(&mut self) -> Result<String, ParseError>

pub fn parse_path(&mut self) -> Result<IdentPath, ParseError>

Parse a ‘path’ (I.e. std::math::powf) from the current cursor position, into a path segment representation.

Notes

Always expects atleast one identifier, can optionally start with ‘::’, to force the path to reference absolutely from the ‘base’.

pub fn parse_spanned_path(&mut self) -> Result<SpannedIdentPath, ParseError>

Parse a ‘path’ (I.e. std::math::powf) from the current cursor position, into a path segment representation, also returns the span of the parsed path.

Notes

Always expects atleast one identifier, can optionally start with ‘::’, to force the path to reference absolutely from the ‘base’.

pub fn parse_ref_and_refmut(&mut self) -> Result<RefType, ParseError>

Parses a & and &mut

pub fn parse_array_or_slice_ty(&mut self) -> Result<Ty, ParseError>

Parse an array or slice type from the current cursor position.

Example

[Ty; N]
[Ty]

Notes

Must be verified the type being parsed is NOT a unit type () by the caller.

pub fn parse_tuple_ty(&mut self) -> Result<Ty, ParseError>

Parse a tuple type from the current cursor position.

Example

(Ty, Ty, ...)

Notes

Must be verified the type being parsed is NOT a unit type () by the caller.

pub fn parse_ty(&mut self) -> Result<Ty, ParseError>

Parse a Ty from the current cursor position.

Notes

Always expects atleast one Identifier.

pub fn parse_impl_ty_ident(&mut self) -> Result<SpannedIdentPath, ParseError>

Parse the type identifier for an impl block.

pub fn parse_visibility(&mut self) -> Result<Visibility, ParseError>

Checks if the current keyword is pub, if it is, consume it and return Ok(Visibility::Public), otherwise Ok(Visibility::Private), If there are no tokens, return Err.

pub fn parse_mutability(&mut self) -> Result<Mutability, ParseError>

Checks if the current keyword is mut, if it is, consume it and return Ok(Mutability::Mutable), otherwise Ok(Mutability::Immutable), If there are no tokens, return Err.

pub fn parse_root(&mut self) -> Result<ASTRoot, ParseError>

Parse the ‘root’ AST, starting from the beginning of the file. This is the function you want to call to fully parse a token stream.

impl Parser<'_, '_>

fn check_kind<T: Into<TokenKind>>(&self, expected_kind: T) -> bool

Peek at the current token, return true if the TokenKind of the Token matches the expected_kind

fn check_kind_at_err<T: Into<TokenKind>>( &self, offset: u32, expected_kind: T, ) -> Result<bool, ParseError>

Peek at the specified offset, return true if the TokenKind of the Token matches the expected_kind

fn check_kind_err<T: Into<TokenKind>>( &self, expected_kind: T, ) -> Result<bool, ParseError>

Peek at the current token, return true if the TokenKind of the Token matches the expected_kind

fn check_kind_at<T: Into<TokenKind>>( &self, offset: u32, expected_kind: T, ) -> bool

Peek at the specified offset, return true if the TokenKind of the Token matches the expected_kind

fn check_kind_advance<T: Into<TokenKind>>(&mut self, expected_kind: T) -> bool

Peek at the current token, return true and advance the cursor if the TokenKind of the Token matches the expected_kind

fn check_kind_advance_err<T: Into<TokenKind>>( &mut self, expected_kind: T, ) -> Result<bool, ParseError>

Peek at the current token, return true and advance the cursor if the TokenKind of the Token matches the expected_kind

fn check_ident(&self) -> bool

Peek at the current token, return true if the TokenKind of the Token is an Identifier.

fn check_ident_at(&self, offset: u32) -> bool

Peek at an offset from the current token, return true if the TokenKind of the Token is an Identifier.

fn check_ident_advance(&mut self) -> Option<String>

Peek at the current token, check if the TokenKind of the Token is an Identifier.

Returns

Some(ident_value) and advance the cursor if it is an Identifier, else None.

fn get_punctuation_sequence<const N: usize>( &self, offset: u32, ) -> Option<[Punctuation; N]>

Get a sequence of Punctuation tokens of length N from a specified offset from the current cursor position.

Returns

Some(arr) if the next N tokens from the specified offset are all Punctuation token kinds.

None if any of the next N tokens from the specified offset are not of kind Punctuation.

impl Parser<'_, '_>

fn expect_literal(&mut self) -> Result<LiteralKind, ParseError>

Expect that the token at the current cursor position is a literal.

This function will advance the cursor by one position if a literal is found.

Returns

Ok(()) if the token at the cursor position is a literal.

Err(ParseErrorKind::ExpectedLiteral) if the token found at the current cursor position was not a literal.

Err(ParseErrorKind::ExpectedTokenFoundNone) if there are no more tokens in the stream, but it was expected that there were.

fn expect_kind<T: Into<TokenKind>>( &mut self, expected_kind: T, ) -> Result<(), ParseError>

Expect that the token at the current cursor position is of a specified TokenKind.

This function will advance the cursor by one position if the expected_kind is found.

Returns

Ok(()) if the token at the cursor position is of the expected_kind.

Err(ParseErrorKind::ExpectedToken) if the token found at the current cursor position does not match the expected_kind.

Err(ParseErrorKind::ExpectedTokenFoundNone) if there are no more tokens in the stream, but it was expected that there were.

fn expect_ident(&mut self) -> Result<String, ParseError>

Expect that the token at the current cursor position is an Identifier.

This function will advance the cursor by one position if an Identifier is found.

Returns

Ok(ident_str) if the token at the cursor position is an Identifier.

Err(ParseErrorKind::ExpectedIdentifier) if the token found at the current cursor position was not an Identifier.

Err(ParseErrorKind::ExpectedIdentifierFoundNone) if there are no more tokens in the stream, but it was expected that there were.

fn expect_ident_spanned(&mut self) -> Result<SpannedIdent, ParseError>

Expect that the token at the current cursor position is an Identifier.

This function will advance the cursor by one position if an Identifier is found.

Returns

Ok((ident_str, span)) if the token at the cursor position is an Identifier.

Err(ParseErrorKind::ExpectedIdentifier) if the token found at the current cursor position was not an Identifier.

Err(ParseErrorKind::ExpectedIdentifierFoundNone) if there are no more tokens in the stream, but it was expected that there were.

impl Parser<'_, '_>

pub fn begin_span(&self) -> u32

Start a span at the beginning of the current token. Shorthand for self.cursor.get_current_source_start.

pub fn finish_span(&self, begin: u32) -> SourceSpan

Construct a span by using a supplied beginning position, and the end of the previous token as the end position. Shorthand for self.cursor.get_current_source_start.

impl Parser<'_, '_>

fn check_token_invalid(token: &Token) -> Result<(), ParseError>

Checks if the token is valid, returning an Err if it isn’t.

Returns

Ok(()) if the token is valid,

Err(ParseError) if the token is invalid.

fn find_next_significant_token( &self, ) -> Result<Option<(u32, &Token)>, ParseError>

Find the next significant token, I.e. A token that isn’t mut or something similar, and that may identify which variant we should try to parse.

Returns

Ok(Some(offset_from_current_token, significant_token)) if there is no error found while checking tokens.

Ok(None) if no significant token was found, but there were no errors either.

Err(ParseError) if an invalid token was found while checking next tokens.

fn find_next_significant_token_with_offset( &self, offset: u32, ) -> Result<Option<(u32, &Token)>, ParseError>

Find the next significant token, I.e. A token that isn’t mut or something similar, and that may identify which variant we should try to parse, starting from a beginning offset from the current cursor position offset.

Returns

Ok(Some(offset_from_current_token, significant_token)) if there is no error found while checking tokens.

Ok(None) if no significant token was found, but there were no errors either.

Err(ParseError) if an invalid token was found while checking next tokens.

fn expect_next_significant_token_with_offset( &self, offset: u32, ) -> Result<(u32, &Token), ParseError>

Find the next significant token, I.e. A token that isn’t mut or something similar, and that may identify which variant we should try to parse, starting from a beginning offset from the current cursor position offset.

Returns

Ok(offset_from_current_token, significant_token) if there is no error found while checking tokens.

Err(ParseError) if an invalid token was found while checking next tokens, or if there were no more tokens available.

fn expect_next_significant_token(&self) -> Result<(u32, &Token), ParseError>

Find the next significant token, I.e. A token that isn’t mut or something similar, and that may identify which variant we should try to parse.

Returns

Ok(offset_from_current_token, significant_token) if there is no error found while checking tokens.

Err(ParseError) if an invalid token was found while checking next tokens, or if there were no more tokens available.

impl Parser<'_, '_>

fn parse_block_like_no_delimiter<T, F: FnMut(&mut Self) -> Result<T, ParseError>>( &mut self, close: Punctuation, f: F, ) -> Result<Vec<T>, ParseError>

Parses a ‘block’-like pattern, where the expected token stream to parse resembles:

<OpenPunctuation>
    <Element>+
<ClosePunctuation>

Notes

This function expects that the <OpenPunctuation> has already been consumed and that the cursor is at the start of the first element in the block.

This function allows for empty blocks.

Generic Parameters

T is the type of the element the block will parse. F is a function that will actually do the parsing of each element, and the function is passed self.

fn parse_block_like_impl<const TRAIL_DELIM: bool, T, F: FnMut(&mut Self) -> Result<T, ParseError>>( &mut self, delimiter: Punctuation, close: Punctuation, f: F, ) -> Result<Vec<T>, ParseError>

Parses a ‘block’-like pattern, where the expected token stream to parse resembles:

<OpenPunctuation>
    <Element><Delimiter>
    <Element><Delimiter>?
<ClosePunctuation>

Notes

This function expects that the <OpenPunctuation> has already been consumed and that the cursor is at the start of the first element in the block.

This function allows for empty blocks.

Generic Parameters

If TRAIL_DELIM is set true, the algorithm will allow a trailing delimiter after the last element but before the closing Punctuation, otherwise this will cause an error. T is the type of the element the block will parse. F is a function that will actually do the parsing of each element, and the function is passed self.

fn parse_block_like<T, F: FnMut(&mut Self) -> Result<T, ParseError>>( &mut self, delimiter: Punctuation, close: Punctuation, f: F, ) -> Result<Vec<T>, ParseError>

Parses a ‘block’-like pattern, where the expected token stream to parse resembles:

<OpenPunctuation>
    <Element><Delimiter>
    <Element><Delimiter>?
<ClosePunctuation>

Notes

This function expects that the <OpenPunctuation> has already been consumed and that the cursor is at the start of the first element in the block.

This function allows for empty blocks.

This function allows for trailing delimiters on the last element before the end of the block.

Generic Parameters

T is the type of the element the block will parse. F is a function that will actually do the parsing of each element, and the function is passed self.

fn parse_block_like_no_trail<T, F: FnMut(&mut Self) -> Result<T, ParseError>>( &mut self, delimiter: Punctuation, close: Punctuation, f: F, ) -> Result<Vec<T>, ParseError>

Parses a ‘block’-like pattern, where the expected token stream to parse resembles:

<OpenPunctuation>
    <Element><Delimiter>
    <Element><Delimiter>?
<ClosePunctuation>

Notes

This function expects that the <OpenPunctuation> has already been consumed and that the cursor is at the start of the first element in the block.

This function allows for empty blocks.

This function does not allow for trailing delimiters on the last element before the end of the block.

Generic Parameters

T is the type of the element the block will parse. F is a function that will actually do the parsing of each element, and the function is passed self.

impl Parser<'_, '_>

fn make_error(&self, kind: ParseErrorKind) -> ParseError

Returns an error with the specified error kind, that is for the token currently pointed to by the cursor.

fn no_token_error(&self) -> ParseError

Returns an error representing that the token stream ended unexpectedly.

fn peek_at(&self, offset: u32) -> Result<&Token, ParseError>

Returns the token at the offset, or ParseErrorKind::NoTokens otherwise.

fn peek(&self) -> Result<&Token, ParseError>

Returns the token at the current cursor position, or ParseErrorKind::NoTokens otherwise.