kitlang

Module prelude

Public re-exports of commonly used items from the Kitlang compiler crate.

Re-exports

pub use crate::lexer::tokenise;

pub use crate::lexer::tokenise_stripped;

pub use crate::token::Keyword;

pub use crate::token::Punctuation;

pub use crate::token::Token;

pub use crate::token::TokenKind;

pub use crate::compiler::Compiler;

pub use crate::intermediate::hir::ProgramMetaData;

pub use crate::intermediate::mir::MIR;

pub use crate::KitlangError;

pub use crate::KitlangResult;

pub use crate::execute_source_string;

pub use crate::parse_source_string_to_mir;

Macros

debug

Logs a message at the debug level.

error

Logs a message at the error level.

info

Logs a message at the info level.

log

The standard logging macro.

log_enabled

Determines if a message logged at the specified level in that module will be logged.

register_native_fn

trace

Logs a message at the trace level.

warn

Logs a message at the warn level.

Structs

Metadata

Metadata about a log message.

MetadataBuilder

Builder for Metadata.

ParseLevelError

The type returned by from_str when the string doesn’t match any of the log levels.

Record

The “payload” of a log message.

RecordBuilder

Builder for Record.

SetLoggerError

The type returned by set_logger if set_logger has already been called.

Enums

Level

An enum representing the available verbosity levels of the logger.

LevelFilter

An enum representing the available verbosity level filters of the logger.

Constants

STATIC_MAX_LEVEL

The statically resolved maximum log level.

Traits

Log

A trait encapsulating the operations required of a logger.

Functions

logger

Returns a reference to the logger.

max_level

Returns the current maximum log level.

set_boxed_logger

Sets the global logger to a Box<Log>.

set_logger

Sets the global logger to a &'static Log.

set_logger_racy⚠

A thread-unsafe version of set_logger.

set_max_level

Sets the global maximum log level.

set_max_level_racy⚠

A thread-unsafe version of set_max_level.