kitlang

Module spanned_error

Provides utilities for building formatted error messages with source code context and highlights.

This module is used throughout the compiler to format different errors that may occur, that include source code snippets and highlighting of the exact location of the error, along with messages that describe the error in detail.

Structs

SpannedErrorBuilder

Used for building pretty formatted error messages based on an error span and the source code string that was used when this error originated.

SpannedErrorLine

Represents a single line in the the “source” view of an error message output. Each source code line can have zero or more ‘comment’ lines, that can provide more information or highlight the problem area.

Constants

highlight_color 🔒

prefix_color 🔒

Functions

count_new_lines 🔒

Counts the number of newline (\n) characters in a string slice.

get_lines 🔒

Converts a span and a source code string, into a Vec of SpannedErrorLines. This finds all lines that the provided SourceSpan touches. Along with filling out meta-data about the line such as it’s line number and where the line starts within the provided source_string.

get_segments_from_source 🔒

Get the 3 segments from the source string and self.

substr_safe 🔒

Safely substring a string, clamping the range to the valid values.

substr_safe_no_trailing_return 🔒

Safely substring a string, clamping the range to the valid values. If the resulting clamped substring ends with ‘\r’, it is removed.