kitlang::spanned_error

Function get_lines

fn get_lines(source_string: &str, span: SourceSpan) -> Vec<SpannedErrorLine>

Converts a span and a source code string, into a Vec of SpannedErrorLines. This finds all lines that the provided SourceSpan touches. Along with filling out meta-data about the line such as it’s line number and where the line starts within the provided source_string.