kitlang

Module token

Defines the type of tokens and keywords used in the Kitlang compiler.

It provides TokenKind for representing different categories of tokens (keywords, punctuation, literals, etc), and the final Token struct which encapsulates a token’s kind along with its position via a start and end character in the source code used for error handling and diagnostic formatting.

Keywords and Punctuations are defined via macros to avoid code duplication and to allow for easy extension in the future if required. These macros define the string or character representations of each keyword/punctuation along with methods for parsing from strings/chars and converting back to strings/chars for parsing and displaying.

Macros

define_keywords 🔒

define_punctuation 🔒

Structs

Token

A lexical token produced by the lexer.

Enums

Keyword

LiteralKind

Punctuation

TokenKind