kitlang::compiler

Struct CompilerContext

pub struct CompilerContext {
    source: String,
    pub meta: ProgramMetaData,
    pub no_stdlib: bool,
}

A context structure for the compiler, holds state relevant to the compilation process.

Fields
source: String

The source code being compiled.

meta: ProgramMetaData

Metadata about the program being compiled.

no_stdlib: bool

Whether to compile without the standard library.

Implementations

impl CompilerContext

pub fn new(source: String) -> Self

Create a new compiler context with the given source code and metadata.

impl CompilerContext

pub fn meta(&self) -> &ProgramMetaData

Get a reference to the program’s meta data.

pub fn meta_mut(&mut self) -> &mut ProgramMetaData

Get a mutable reference to the program’s namespace.

pub fn namespace(&self) -> &RootNamespace

Get a reference to the program’s namespace.

pub fn namespace_mut(&mut self) -> &mut RootNamespace

Get a mutable reference to the program’s namespace.

pub fn type_registry(&self) -> &TypeRegistry

Get a reference to the program’s type registry.

pub fn type_registry_mut(&mut self) -> &mut TypeRegistry

Get a mutable reference to the program’s type registry.

pub fn type_map(&self) -> &TypeMap

Get a reference to the program’s type registry.

pub fn type_map_mut(&mut self) -> &mut TypeMap

Get a mutable reference to the program’s type registry.

impl CompilerContext

const CORE_LIB: &str = "// @lang\nimpl i8 {\n pub fn to_string(self) -> string { i64_to_string(self as i64) }\n pub fn from_string(s: string) -> (Self, bool) {\n let result = string_to_i64(s);\n (result.0 as Self, result.1)\n }\n pub fn from_string_or(s: string, default: Self) -> Self {\n let (v, err) = i8::from_string(s);\n if err { default } else { v }\n }\n\n pub fn sqrt(self) -> Self { i64_sqrt(self as i64) as Self }\n pub fn abs(self) -> Self { i64_abs(self as i64) as Self }\n}\n// @lang\nimpl i16 {\n pub fn to_string(self) -> string { i64_to_string(self as i64) }\n pub fn from_string(s: string) -> (Self, bool) {\n let result = string_to_i64(s);\n (result.0 as Self, result.1)\n }\n pub fn from_string_or(s: string, default: Self) -> Self {\n let (v, err) = i16::from_string(s);\n if err { default } else { v }\n }\n\n pub fn sqrt(self) -> Self { i64_sqrt(self as i64) as Self }\n pub fn abs(self) -> Self { i64_abs(self as i64) as Self }\n}\n// @lang\nimpl i32 {\n pub fn to_string(self) -> string { i64_to_string(self as i64) }\n pub fn from_string(s: string) -> (Self, bool) {\n let result = string_to_i64(s);\n (result.0 as Self, result.1)\n }\n pub fn from_string_or(s: string, default: Self) -> Self {\n let (v, err) = i32::from_string(s);\n if err { default } else { v }\n }\n\n pub fn sqrt(self) -> Self { i64_sqrt(self as i64) as Self }\n pub fn abs(self) -> Self { i64_abs(self as i64) as Self }\n}\n// @lang\nimpl i64 {\n pub fn to_string(self) -> string { i64_to_string(self) }\n pub fn from_string(s: string) -> (Self, bool) { string_to_i64(s) }\n pub fn from_string_or(s: string, default: Self) -> Self {\n let (v, err) = i64::from_string(s);\n if err { default } else { v }\n }\n\n pub fn sqrt(self) -> Self { i64_sqrt(self) }\n pub fn abs(self) -> Self { i64_abs(self) }\n}\n// @lang\nimpl i128 {\n pub fn to_string(self) -> string { i64_to_string(self as i64) }\n pub fn from_string(s: string) -> (Self, bool) {\n let result = string_to_i64(s);\n (result.0 as Self, result.1)\n }\n pub fn from_string_or(s: string, default: Self) -> Self {\n let (v, err) = i128::from_string(s);\n if err { default } else { v }\n }\n\n pub fn sqrt(self) -> Self { i64_sqrt(self as i64) as Self }\n pub fn abs(self) -> Self { i64_abs(self as i64) as Self }\n}\n// @lang\nimpl isize {\n pub fn to_string(self) -> string { i64_to_string(self as i64) }\n pub fn from_string(s: string) -> (Self, bool) {\n let result = string_to_i64(s);\n (result.0 as Self, result.1)\n }\n pub fn from_string_or(s: string, default: Self) -> Self {\n let (v, err) = isize::from_string(s);\n if err { default } else { v }\n }\n\n pub fn sqrt(self) -> Self { i64_sqrt(self as i64) as Self }\n pub fn abs(self) -> Self { i64_abs(self as i64) as Self }\n}\n\n// @lang\nimpl u8 {\n pub fn to_string(self) -> string { u64_to_string(self as u64) }\n pub fn from_string(s: string) -> (Self, bool) {\n let result = string_to_u64(s);\n (result.0 as Self, result.1)\n }\n pub fn from_string_or(s: string, default: Self) -> Self {\n let (v, err) = u8::from_string(s);\n if err { default } else { v }\n }\n\n pub fn sqrt(self) -> Self { u64_sqrt(self as u64) as Self }\n}\n// @lang\nimpl u16 {\n pub fn to_string(self) -> string { u64_to_string(self as u64) }\n pub fn from_string(s: string) -> (Self, bool) {\n let result = string_to_u64(s);\n (result.0 as Self, result.1)\n }\n pub fn from_string_or(s: string, default: Self) -> Self {\n let (v, err) = u16::from_string(s);\n if err { default } else { v }\n }\n\n pub fn sqrt(self) -> Self { u64_sqrt(self as u64) as Self }\n}\n// @lang\nimpl u32 {\n pub fn to_string(self) -> string { u64_to_string(self as u64) }\n pub fn from_string(s: string) -> (Self, bool) {\n let result = string_to_u64(s);\n (result.0 as Self, result.1)\n }\n pub fn from_string_or(s: string, default: Self) -> Self {\n let (v, err) = u32::from_string(s);\n if err { default } else { v }\n }\n\n pub fn sqrt(self) -> Self { u64_sqrt(self as u64) as Self }\n}\n// @lang\nimpl u64 {\n pub fn to_string(self) -> string { u64_to_string(self) }\n pub fn from_string(s: string) -> (Self, bool) { string_to_u64(s) }\n pub fn from_string_or(s: string, default: Self) -> Self {\n let (v, err) = u64::from_string(s);\n if err { default } else { v }\n }\n\n pub fn sqrt(self) -> Self { u64_sqrt(self) }\n}\n// @lang\nimpl u128 {\n pub fn to_string(self) -> string { u64_to_string(self as u64) }\n pub fn from_string(s: string) -> (Self, bool) {\n let result = string_to_u64(s);\n (result.0 as Self, result.1)\n }\n pub fn from_string_or(s: string, default: Self) -> Self {\n let (v, err) = u128::from_string(s);\n if err { default } else { v }\n }\n\n pub fn sqrt(self) -> Self { u64_sqrt(self as u64) as Self }\n}\n// @lang\nimpl usize {\n pub fn to_string(self) -> string { u64_to_string(self as u64) }\n pub fn from_string(s: string) -> (Self, bool) {\n let result = string_to_u64(s);\n (result.0 as Self, result.1)\n }\n pub fn from_string_or(s: string, default: Self) -> Self {\n let (v, err) = usize::from_string(s);\n if err { default } else { v }\n }\n\n pub fn sqrt(self) -> Self { u64_sqrt(self as u64) as Self }\n}\n\n// @lang\nimpl f32 {\n pub fn to_string(self) -> string { f64_to_string(self as f64) }\n pub fn from_string(s: string) -> (Self, bool) {\n let result = string_to_f64(s);\n (result.0 as Self, result.1)\n }\n pub fn from_string_or(s: string, default: Self) -> Self {\n let (v, err) = f32::from_string(s);\n if err { default } else { v }\n }\n\n pub fn sqrt(self) -> Self { f64_sqrt(self as f64) as Self }\n pub fn abs(self) -> Self { f64_abs(self as f64) as Self }\n}\n// @lang\nimpl f64 {\n pub fn to_string(self) -> string { f64_to_string(self) }\n pub fn from_string(s: string) -> (Self, bool) { string_to_f64(s) }\n pub fn from_string_or(s: string, default: Self) -> Self {\n let (v, err) = f64::from_string(s);\n if err { default } else { v }\n }\n\n pub fn sqrt(self) -> Self { f64_sqrt(self) }\n pub fn abs(self) -> Self { f64_abs(self) }\n}\n\n// @lang\nimpl bool {\n pub fn to_string(self) -> string { bool_to_string(self) }\n pub fn from_string(s: string) -> (Self, bool) { string_to_bool(s) }\n pub fn from_string_or(s: string, default: Self) -> Self {\n let (v, err) = bool::from_string(s);\n if err { default } else { v }\n }\n}\n\n// @lang\nimpl string {\n pub fn to_string(self) -> string { self }\n pub fn from_string(s: string) -> string { s }\n\n pub fn len(self) -> usize { str_len(self) as usize }\n pub fn is_empty(self) -> bool { self.len() == 0 as usize }\n pub fn to_lower(self) -> string { to_lower(self) }\n}\n\n// @lang\nimpl () {\n pub fn to_string(self) -> string { \"()\" }\n pub fn from_string(s: string) { }\n}"

const INTRINSICS_LIB: &str = "native global fn to_lower(s: string) -> string;\nnative global fn str_len(s: string) -> usize;\n\nnative global fn print(s: string);\nnative global fn println(s: string);\n\nnative global fn read_line() -> string;\nnative global fn input(s: string) -> string;\nnative global fn input_placeholder(s: string, default: string) -> string;\n\nnative global fn i64_to_string(i: i64) -> string;\nnative global fn string_to_i64(s: string) -> (i64, bool);\nnative global fn u64_to_string(u: u64) -> string;\nnative global fn string_to_u64(s: string) -> (u64, bool);\nnative global fn f64_to_string(f: f64) -> string;\nnative global fn string_to_f64(s: string) -> (f64, bool);\nnative global fn bool_to_string(f: bool) -> string;\nnative global fn string_to_bool(s: string) -> (bool, bool);\n\nnative global fn i64_sqrt(i: i64) -> i64;\nnative global fn u64_sqrt(u: u64) -> u64;\nnative global fn f64_sqrt(f: f64) -> f64;\n\nnative global fn i64_abs(i: i64) -> i64;\nnative global fn f64_abs(f: f64) -> f64;"

pub fn source(&self) -> String

Get the full source code, with standard library injected if applicable.

pub fn raw_source(&self) -> &str

Get the raw source code, without standard library injected, regardless of settings.