kitlang::intermediate::mir::lowerer

Struct HIRToMIRFuncLowerer

struct HIRToMIRFuncLowerer<'a> {
    pub program_meta_data: &'a ProgramMetaData,
    pub body: Body,
    pub func_owner_id: OwnerDefId,
    pub func_body_id: HirId,
    pub lut: HashMap<HirId, LocalId>,
    pub block_stack: Vec<HIRToMIRBlockBuilder>,
    pub state: HIRToMIRFuncLowererState,
    pub errors: Vec<LoweringError>,
}

Fields
program_meta_data: &'a ProgramMetaDatabody: Bodyfunc_owner_id: OwnerDefIdfunc_body_id: HirIdlut: HashMap<HirId, LocalId>block_stack: Vec<HIRToMIRBlockBuilder>state: HIRToMIRFuncLowererStateerrors: Vec<LoweringError>
Implementations

impl<'a> HIRToMIRFuncLowerer<'a>

pub fn from_func_id( hlir: &HLIR, meta_data: &'a ProgramMetaData, func_id: OwnerDefId, ) -> Result<Body, Vec<LoweringError>>

impl HIRToMIRFuncLowerer<'_>

const DEBUG_BLOCK_CREATION: bool = false

fn create_block_builder(&mut self)

fn builder(&self) -> Option<&HIRToMIRBlockBuilder>

fn builder_mut(&mut self) -> Option<&mut HIRToMIRBlockBuilder>

fn builder_expect(&self) -> &HIRToMIRBlockBuilder

fn builder_mut_expect(&mut self) -> &mut HIRToMIRBlockBuilder

fn emit_block(&mut self) -> Option<BasicBlockId>

fn emit_and_replace_block(&mut self) -> Option<BasicBlockId>

fn emit_final_block(&mut self) -> Option<BasicBlockId>

impl HIRToMIRFuncLowerer<'_>

fn is_directive_set(&self) -> bool

fn process_statement_expr( &mut self, _statement: &Statement, hlir: &HLIR, expr_id: HirId, )

fn get_mutability_of_local(&self, local_id: LocalId) -> Mutability

fn new_temp_local_with_mut(&mut self, mutable: Mutability) -> LocalId

fn new_temp_local(&mut self) -> LocalId

const fn last_local(&self) -> LocalId

fn last_target(&mut self) -> AssignTarget

fn handle_resolved_id(&mut self, resolved: ResolvedID, hlir: &HLIR)

fn visit_expr_assigned( &mut self, expr_id: HirId, hlir: &HLIR, ) -> Option<AssignTarget>

fn visit_expr_expect_owner( &mut self, expr_id: HirId, hlir: &HLIR, ) -> Option<OwnerDefId>

impl HIRToMIRFuncLowerer<'_>

fn update_branch_else_target( &mut self, branch_id: BasicBlockId, new_target: BasicBlockId, ) -> bool

Updates the else target of a branch block.

Returns

true if the target was updated, false otherwise.

fn update_branch_targets( &mut self, branch_id: BasicBlockId, new_true_target: BasicBlockId, new_false_target: BasicBlockId, ) -> bool

Updates the true and else targets of a branch block.

Returns

true if the target was updated, false otherwise.

fn update_goto_target( &mut self, block_id: BasicBlockId, new_target: BasicBlockId, ) -> bool

Updates the goto target of a goto block.

Returns

true if the target was updated, false otherwise.

fn update_goto_targets_from_loop( &mut self, hlir: &HLIR, expr_id: HirId, block_id: BasicBlockId, )

Updates all break goto targets from the current loop to the given block.

Parameters

expr_id: The expression ID of the full loop expression used for error reporting.

block_id: The block ID to update the break targets to.

fn block_target_to_rvalue(target: Option<AssignTarget>) -> RValue

Converts a block target to an RValue.

impl HIRToMIRFuncLowerer<'_>

fn lower_if( &mut self, hlir: &HLIR, if_expr_id: HirId, condition: HirId, true_block: HirId, else_expr: Option<&HirId>, )

fn lower_loop(&mut self, hlir: &HLIR, loop_expr_id: HirId, loop_block_id: HirId)

fn lower_for( &mut self, hlir: &HLIR, for_expr_id: HirId, binding_id: HirId, iterable_id: HirId, block_id: HirId, )

fn lower_while( &mut self, hlir: &HLIR, while_expr_id: HirId, condition: HirId, block: HirId, )

fn lower_call( &mut self, hlir: &HLIR, expr_id: HirId, call_id: HirId, args: &[HirId], )

fn lower_method_call( &mut self, hlir: &HLIR, expr_id: HirId, target_id: HirId, method_name: &str, args: &[HirId], )

fn lower_field_access( &mut self, hlir: &HLIR, expr_id: HirId, target_id: HirId, field_name: &str, )

fn lower_struct_init( &mut self, hlir: &HLIR, expr_id: HirId, struct_init: &StructInitialisation, )